BivMetaAnalysis {INLA}	R Documentation

Bivariate Meta Analysis

Description

Data are taken from a meta-analysis to compare the utility of three types of diagnostic imaging - lymphangiography (LAG), computed tomography (CT) and magnetic resonance (MR) - to detect lymph node metastases in patients with cervical cancer. The dataset consists of a total of 46 studies: the first 17 for LAG, the following 19 for CT and the last 10 for MR.

Usage

BivMetaAnalysis

Format

A data frame with 92 observations on the following 9 variables.

N

a numeric vector

Y

a numeric vector

diid

a numeric vector

lag.tp

a numeric vector

lag.tn

a numeric vector

ct.tp

a numeric vector

ct.tn

a numeric vector

mr.tp

a numeric vector

mr.tn

a numeric vector

References

J. Scheidler and H. Hricak and K. K. Yu and L. Subak and M. R. Segal,"Radiological evaluation of lymph node metastases in patients with cervical cancer: a meta-analysis",JAMA 1997

Examples

data(BivMetaAnalysis)

[Package INLA version 22.02.16-2 Index]