Drivers {INLA}	R Documentation

Time series with seasonal effect

Description

Montly total of car drivers killed or several injuried in England from January 1969 to December 1984

Format

A data frame with 204 observations on the following 4 variables.

y

Number of deaths

belt

Indicator of weather the belt was compulsory to use (1) or not (0)

trend

time (in months)

seasonal

time (in months)

Details

NB: The last 12 lines of the data set have the first column set to NULL since these data where not observed but we want to predict them.

References

Rue, H and Held, L. (2005) Gaussian Markov Random Fields - Theory and Applications Chapman and Hall

Examples

data(Drivers)

[Package INLA version 22.02.16-2 Index]