Germany {INLA}	R Documentation

Disease Mapping

Description

Cases of Oral cavity cancer in Germany from 1986-1990

Format

A data frame with 544 observations on the following 4 variables.

region

Region of Germany

E

Fixed quantity which accounts for number of people in the district (offset)

Y

Number of cases

x

covariate measuring smoking consumption

References

Rue, H and Held, L. (2005) Gaussian Markov Random Fields - Theory and Applications Chapman and Hall

Examples

data(Germany)

[Package INLA version 22.02.16-2 Index]