INLA-package {INLA}	R Documentation

Integrated Nested Laplace Approximation

Description

Package to perform full Bayesian analysis on generalised additive mixed models using Integrated Nested Laplace Approximations.

Details

Package

INLA

Version

Currently, this package uses a YY.MM.DD versioning system, and is in heavy development. See https://github.com/hrue/r-inla/ and https://www.r-inla.org

License

GPL2

See the web-site https://www.r-inla.org for further details.

Author(s)

Havard Rue, Sara Martino, Finn Lindgren, Daniel Simpson and Andrea Riebler

[Package INLA version 22.02.16-2 Index]