Salm {INLA}	R Documentation

Extra-Poisson variation in dose-response study

Description

Breslow (1984) analyses some mutagenicity assay data (shown below) on salmonella in which three plates have been processed at each dose i of quinoline and the number of revertant colonies of TA98 Salmonella measured

Format

A data frame with 18 observations on the following 3 variables.

y

number of salmonella bacteria

dose

dose of quinoline (mg per plate)

rand

indicator

Source

WinBUGS/OpenBUGS manual Examples VOl.I

Examples

data(Salm)

[Package INLA version 22.02.16-2 Index]