Scotland {INLA}	R Documentation

Conditional Autoregressive (CAR) model for disease mapping

Description

The rate of lip cancer in 56 counties in Scotland is recorder. The data set includes the observed and expected cases (based on the population and its age and sex distribution in the country), a covariate measuring the percentage of the population engaged in agricolture, fishing or forestry and the "position" of each county expressed as a list of adjacent counties

Format

A data frame with 56 observations on the following 4 variables.

Counts

The number of lip cancer registered

E

The expected number of lip cancer

X

The percentage of the population engaged in agricolture, fishing or forestry

Region

The county

Source

OpenBUGS Example manual, GeoBUGS

References

Clayton and Kaldor (1987) and Breslow and Clayton (1993)

Examples

data(Scotland)

[Package INLA version 22.02.16-2 Index]