Seeds {INLA}	R Documentation

Factorial design

Description

Proportion of seeds that germinated on each of 21 plates arranged according to a 2 by 2 factorial layout by seed and type of root extract

Format

A data frame with 21 observations on the following 5 variables.

r

number of germinated seeds per plate

n

number of total seeds per plate

x1

seed type

x2

root extracted

plate

indicator for the plate

Source

WinBUGS/OpenBUGS Manual Example, Vol. I

Examples

data(Seeds)

[Package INLA version 22.02.16-2 Index]