aggregate.gaussian {INLA}	R Documentation

Aggregate Gaussian into an equivalent observation

Description

Aggreagate Gaussians observerd with the same mean and precision, into an eqv triplet, for use with family="agaussian"

Usage

     inla.agaussian(y, s = NULL)
 

Arguments

y
Repeated observations. If y is a matrix, then each row represents repeated observations. if y is a list, then each element of the list is a vector of repeated observations. If y is a vector, then the whole vector represents repeated observations. The optional scaling s, must have the same format as y, ie matrix or vector. NA's in y (and s) are removed and not used or counted. If s is given, then the NA-pattern in y and s must be the same.

s
Optinonal fixed scaling of the precisions. Must be in the same format as y, and have the same NA-pattern. See the documentation for details.

Value

The output is a inla.mdata-object ready for use with family="agaussian". See the example in the documentation.

Author(s)

Havard Rue hrue@r-inla.org

Examples

  A = matrix(1:25,5,5)
  inla.agaussian(A)

  A[1,-1] = NA
  A[2,-(2:3)] = NA
  inla.agaussian(A)
 

[Package INLA version 22.02.16-2 Index]