Install alternative binary builds.

Install a new binary for os unless missing(os), for which the os is chosen interactively among the available builds.

inla.binary.install(
  os = c("CentOS Linux-6", "CentOS Linux-7", "CentOS Linux-8", "CentOS Stream-8",
    "Rocky Linux-8", "Rocky Linux-9", "Manjaro Linux-", "Fedora-33", "Fedora-34",
    "Fedora Linux-35", "Fedora Linux-36", "Fedora Linux-37", "Fedora Linux-38",
    "Fedora Linux-39", "Fedora Linux-40", "Fedora Linux-41", "Ubuntu-16.04",
    "Ubuntu-18.04", "Ubuntu-20.04", "Ubuntu-22.04", "Ubuntu-24.04"),
  path = NULL,
  verbose = TRUE,
  md5.check = TRUE,
  secure.http = TRUE
)

Arguments

os

If os is given, install binary build for this os. If os is not given, chose os interactively among available builds.

path

character. The install path. If NULL the path is derived from INLA package

verbose

Logical. Verbose output if TRUE

md5.check

Logical. If TRUE, stop if md5-checksum-file is not present or md5-checksum fail. If FALSE, ignore md5-checksum check.

secure.http

Logical. Use secure http (ie https://) or http://

Value

Return TRUE if installation was sucessful and FALSE if not.

Author

Havard Rue hrue@r-inla.org

Examples

   if (FALSE) { # \dontrun{
     inla.binary.install()
     inla.binary.install(os = "CentOS Linux-7")
     inla.binary.install(os = "CentOS Linux-7",  path = "~/local/bin/inla.binary")
   } # }