inla.collect.results {INLA}	R Documentation

Collect results from a inla-call

Description

inla.collect.results collect results from a inla-call

Usage

 inla.collect.results(
              results.dir,
              debug=FALSE,
              only.hyperparam=FALSE,
              file.log = NULL,
              file.log2 = NULL)

Arguments

results.dir
The directory where the results of the inla run are stored

debug
Logical. If TRUE some debugging information are printed

only.hyperparam
Binary variable indicating wheather only the results for the hyperparameters should be collected

file.log
Character. The filename, if any, of the logfile for the internal calculations

file.log2
Character. The filename, if any, of the logfile2 for the internal calculations

silent
Internal use only

Details

This function is mainly used inside inla to collect results after running the inla function. It can also be used to collect results into R after having runned a inla section outside R.

Value

The function returns an object of class "inla", see the help file for inla for details.

[Package INLA version 22.02.16-2 Index]