control.bgev.default {INLA}	R Documentation

Control variables in control.bgev.default

Description

Control variables in control.bgev.default for use in inla

Usage

inla.set.control.bgev.default.default(...)
control.bgev.default(beta.ab, q.location, q.mix, q.spread)

Arguments

...
Possible arguments

q.location
The quantile level for the location parameter

q.spread
The quantile level for the spread parameter (must be < 0.5)

q.mix
The lower and upper quantile level for the mixing function

beta.ab
The parameters a and b in the Beta mixing function

Value

The control.bgev-list is set within the corresponding control.family-list as control parameters to the family="bgev" The function control.bgev.default is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.bgev.default.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]