Control variables in control.compute

Control variables in control.compute for use in inla

inla.set.control.compute.default(...)
control.compute(config, cpo, dic, gdensity, graph, hyperpar, mlik, openmp.strategy, po, q, return.marginals, smtp, waic)

Arguments

...

Possible arguments

openmp.strategy

The computational strategy to use: 'small', 'medium', 'large', 'huge' and 'default'. There are also two options for the pardiso solver: 'pardiso.serial' and 'pardiso.parallel'. The difference is how the parallelisation is done, and is tuned for 'small'-sized models, 'medium'-sized models, etc. The default option tries to make an educated guess, but this allows to overide this selection. Default is 'default'

hyperpar

A boolean variable if the marginal for the hyperparameters should be computed. Default TRUE.

return.marginals

A boolean variable if the marginals for the latent field should be returned (although it is computed). Default TRUE

dic

A boolean variable if the DIC-value should be computed. Default FALSE.

mlik

A boolean variable if the marginal likelihood should be computed. Default TRUE.

cpo

A boolean variable if the cross-validated predictive measures (cpo, pit) should be computed (default FALSE)

po

A boolean variable if the predictive ordinate should be computed (default FALSE)

waic

A boolean variable if the Watanabe-Akaike information criteria should be computed (default FALSE)

q

A boolean variable if binary images of the precision matrix, the reordered precision matrix and the Cholesky triangle should be generated. (Default FALSE.)

config

A boolean variable if the internal GMRF approximations be stored. (Default FALSE. EXPERIMENTAL)

smtp

The sparse-matrix solver, one of 'default', 'taucs', 'band' or 'pardiso' (default inla.getoption("smtp"))

graph

A boolean variable if the graph itself should be returned. (Default FALSE.)

gdensity

A boolean variable if the Gaussian-densities itself should be returned. (Default FALSE.)

Value

The function control.compute is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.compute.default returns a list with all the default values of all parameters within this control statement.

See also