control.fixed {INLA}	R Documentation

Control variables in control.fixed

Description

Control variables in control.fixed for use in inla

Usage

inla.set.control.fixed.default(...)
control.fixed(cdf, compute, correlation.matrix, expand.factor.strategy, mean, mean.intercept, prec, prec.intercept, quantiles, remove.names)

Arguments

...
Possible arguments

cdf
A list of values to compute the CDF for, for all fixed effects

quantiles
A list of quantiles to compute for all fixed effects

expand.factor.strategy
The strategy used to expand factors into fixed effects based on their levels. The default strategy is us use the model.matrix-function for which NA's are not allowed (expand.factor.strategy="model.matrix") and levels are possible removed. The alternative option (expand.factor.strategy="inla") use an inla-spesific expansion which expand a factor into one fixed effects for each level, do allow for NA's and all levels are present in the model. In this case, factors MUST BE factors in the data.frame/list and NOT added as .+factor(x1)+. in the formula only.

mean
Prior mean for all fixed effects except the intercept. Alternatively, a named list with specific means where name=default applies to unmatched names. For example control.fixed=list(mean=list(a=1, b=2, default=0)) assign 'mean=1' to fixed effect 'a' , 'mean=2' to effect 'b' and 'mean=0' to all others. (default 0.0)

mean.intercept
Prior mean for the intercept (default 0.0)

prec
Default precision for all fixed effects except the intercept. Alternatively, a named list with specific means where name=default applies to unmatched names. For example control.fixed=list(prec=list(a=1, b=2, default=0.01)) assign 'prec=1' to fixed effect 'a' , 'prec=2' to effect 'b' and 'prec=0.01' to all others. (default 0.001)

prec.intercept
Default precision the intercept (default 0.0)

compute
Compute marginals for the fixed effects ? (default TRUE)

correlation.matrix
Compute the posterior correlation matrix for all fixed effects? (default FALSE) OOPS: This option will set up appropriate linear combinations and the results are shown as the posterior correlation matrix of the linear combinations. This option will imply control.inla=list(lincomb.derived.correlation.matrix=TRUE).

remove.names
A vector of names of expanded fixed effects to remove from the model-matrix. This is an expert option, and should only be used if you know what you are doing.

Value

The function control.fixed is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.fixed.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]