Control variables in control.group

Control variables in control.group for use in inla

inla.set.control.group.default(...)
control.group(adjust.for.con.comp, cyclic, fixed, graph, hyper, initial, model, order, param, prior, scale.model)

Arguments

...

Possible arguments

model

Group model (one of 'exchangable', 'exchangablepos', 'ar1', 'ar', 'rw1', 'rw2', 'besag', or 'iid')

order

Defines the order of the model: for model ar this defines the order p, in AR(p). Not used for other models at the time being.

cyclic

Make the group model cyclic? (Only applies to models 'ar1', 'rw1' and 'rw2')

graph

The graph spesification (Only applies to model 'besag')

scale.model

Scale the intrinsic model (RW1, RW2, BESAG) so the generalized variance is 1. (Default TRUE)

adjust.for.con.comp

Adjust for connected components when scale.model=TRUE? (default TRUE)

hyper

Definition of the hyperparameter(s)

initial

(OBSOLETE!) The initial value for the group correlation or precision in the internal scale.

fixed

(OBSOLETE!) A boolean variable if the group correction or precision is assumed to be fixed or random.

prior

(OBSOLETE!) The name of the prior distribution for the group correlation or precision in the internal scale

param

(OBSOLETE!) Prior parameters

Value

The function control.group is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.group.default returns a list with all the default values of all parameters within this control statement.

See also