Control variables in control.hazard

Control variables in control.hazard for use in inla

inla.set.control.hazard.default(...)
control.hazard(constr, cutpoints, diagonal, fixed, hyper, initial, model, n.intervals, param, prior, scale.model, strata.name)

Arguments

...

Possible arguments

model

The model for the baseline hazard model. One of 'rw1' or 'rw2'. (Default 'rw1'.)

hyper

The definition of the hyperparameters.

fixed

(OBSOLETE!) A boolean variable; is the precision for 'model' fixed? (Default FALSE.)

initial

(OBSOLETE!) The initial value for the precision.

prior

(OBSOLETE!) The prior distribution for the precision for 'model'

param

(OBSOLETE!) The parameters in the prior distribution

constr

A boolean variable; shall the 'model' be constrained to sum to zero?

diagonal

An extra constant added to the diagonal of the precision matrix

n.intervals

Number of intervals in the baseline hazard. (Default 15)

cutpoints

The cutpoints to use. If not specified the they are compute from 'n.intervals' and the maximum length of the interval. (Default NULL)

strata.name

The name of the stratefication variable for the baseline hazard in the data.frame

scale.model

Scale the baseline hazard model (RW1, RW2) so the generalized variance is 1. (Default inla.getOption("scale.model.default").)

Value

The function control.hazard is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.hazard.default returns a list with all the default values of all parameters within this control statement.

See also