Control variables in control.link

Control variables in control.link for use in inla

inla.set.control.link.default(...)
control.link(a, fixed, hyper, initial, model, order, param, prior, quantile, variant)

Arguments

...

Possible arguments

model

The name of the link function/model

order

The order of the link function, where the interpretation of order is model-dependent.

variant

The variant of the link function, where the interpretation of variant is model-dependent.

hyper

Definition of the hyperparameter(s) for the link model chosen

quantile

The quantile for quantile link function

a

The parameter a in the LOGa link

initial

(OBSOLETE!) The initial value(s) for the hyperparameter(s)

fixed

(OBSOLETE!) A boolean variable if hyperparmater(s) is/are fixed or random

prior

(OBSOLETE!) The name of the prior distribution(s) for the hyperparmater(s)

param

(OBSOLETE!) The parameters for the prior distribution(s) for the hyperparmater(s)

Value

The control.link-list is set within the corresponding control.family-list as the link is likelihood-familiy spesific. The function control.link is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.link.default returns a list with all the default values of all parameters within this control statement.

See also