control.mix {INLA}	R Documentation

Control variables in control.mix

Description

Control variables in control.mix for use in inla

Usage

inla.set.control.mix.default(...)
control.mix(fixed, hyper, initial, integrator, model, npoints, param, prior)

Arguments

...
Possible arguments

model
The model for the random effect. Currently, only model='gaussian' is implemented

hyper
Definition of the hyperparameter(s) for the random effect model chosen

initial
(OBSOLETE!) The initial value(s) for the hyperparameter(s)

fixed
(OBSOLETE!) A boolean variable if hyperparmater(s) is/are fixed or random

prior
(OBSOLETE!) The name of the prior distribution(s) for the hyperparmater(s)

param
(OBSOLETE!) The parameters for the prior distribution(s) for the hyperparmater(s)

npoints
Number of points used to do the numerical integration (default 101)

integrator
The integration scheme to use (default, quadrature, simpson)

Value

The control.mix -list is set within the corresponding control.family-list a the mixture of the likelihood is likelihood spesific. (This option is EXPERIMENTAL.) The function control.mix is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.mix.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]