control.pardiso

Control variables in control.* for use with inla(). The functions can be used to TAB-complete arguments, and returns a list of the default control arguments, unless overridden by specific input arguments.

control.pardiso(
  verbose = FALSE,
  debug = FALSE,
  parallel.reordering = TRUE,
  nrhs = -1
)

inla.set.control.pardiso.default(...)

Arguments

verbose

Show detailed output (default FALSE)

debug

Show internal debug output (default FALSE)

parallel.reordering

Do reordering in parallel (default TRUE)

nrhs

Number of right-hand sides to solve for in parallel (-1 will determine this adapative)

...

Named arguments passed on to the main function

Details

Extra options controlling the PARDISO library

See also