control.pom {INLA}	R Documentation

Control variables in control.pom

Description

Control variables in control.pom for use in inla

Usage

inla.set.control.pom.default(...)
control.pom(cdf, fast)

Arguments

...
Possible arguments

cdf
character The cdf to use, "logit" (default) or "probit"

fast
Logical Use a faster but approximate form for the probit cdf (default FALSE)?

Value

The control.pom -list is set within the corresponding control.family The function control.pom is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.pom.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]