Improved estimates for the CPO/PIT-values

Improve the estimates of the CPO/PIT-values be recomputing the model-fit by removing data-points.

inla.cpo(
  result,
  force = FALSE,
  mc.cores = NULL,
  verbose = TRUE,
  recompute.mode = TRUE
)

Arguments

result

An object of class inla, ie a result of a call to inla()

force

If TRUE, then recompute all CPO/PIT values and not just those with result$cpo$failure > 0.

mc.cores

The number of cores to use in parallel::mclapply. If is.null(mc.cores), then check getOption("mc.cores") and inla.getOption("num.threads") in that order.

verbose

Run in verbose mode?

recompute.mode

Should be mode (and the integration points) be recomputed when a data-point is removed or not?

Value

The object returned is the same as result but the new improved estimates of the CPO/PIT values replaced.

See also

Author

Havard Rue hrue@r-inla.org

Examples

n = 10
y = rnorm(n)
r = inla(
  y ~ 1,
  data = data.frame(y),
  control.compute = list(cpo=TRUE),
  num.threads = "1:1" # Protect package testing from parallel execution
)
#> Warning: error in running command
#> 
#>  *** inla.core.safe:  The inla program failed, but will rerun in case better initial values may help. try=1/1 
#> Warning: error in running command
#> Error in inla.core.safe(formula = formula, family = family, contrasts = contrasts,     data = data, quantiles = quantiles, E = E, offset = offset,     scale = scale, weights = weights, Ntrials = Ntrials, strata = strata,     lp.scale = lp.scale, link.covariates = link.covariates, verbose = verbose,     lincomb = lincomb, selection = selection, control.compute = control.compute,     control.predictor = control.predictor, control.family = control.family,     control.inla = control.inla, control.fixed = control.fixed,     control.mode = control.mode, control.expert = control.expert,     control.hazard = control.hazard, control.lincomb = control.lincomb,     control.update = control.update, control.lp.scale = control.lp.scale,     control.pardiso = control.pardiso, control.stiles = control.stiles,     control.taucs = control.taucs, control.numa = control.numa,     only.hyperparam = only.hyperparam, inla.call = inla.call,     inla.arg = inla.arg, num.threads = num.threads, keep = keep,     working.directory = working.directory, silent = silent, inla.mode = inla.mode,     safe = FALSE, debug = debug, .parent.frame = .parent.frame): The inla-program exited with an error. Unless you interupted it yourself, please rerun with verbose=TRUE and check the output carefully.
#>   If this does not help, please contact the developers at <help@r-inla.org>.
#> The inla program failed and the maximum number of tries has been reached.
# \donttest{
rr = inla.cpo(
  r,
  force=TRUE,
  mc.cores = 1 # Protect package testing from parallel execution
)
#> Error: object 'r' not found
# }