inla.wkt_is_geocent {INLA}	R Documentation

Handling CRS/WKT

Description

Get and set CRS object or WKT string properties.

Usage

inla.wkt_is_geocent(wkt)

inla.crs_is_geocent(crs)

inla.wkt_get_ellipsoid_radius(wkt)

inla.crs_get_ellipsoid_radius(crs)

inla.wkt_set_ellipsoid_radius(wkt, radius)

inla.crs_set_ellipsoid_radius(crs, radius)

inla.wkt_unit_params()

inla.wkt_get_lengthunit(wkt)

inla.wkt_set_lengthunit(wkt, unit, params = NULL)

inla.crs_get_wkt(crs)

inla.crs_get_lengthunit(crs)

inla.crs_set_lengthunit(crs, unit, params = NULL)

Arguments

wkt
A WKT2 character string

crs
A sp::CRS or inla.CRS object

unit
character, name of a unit. Supported names are "metre", "kilometre", and the aliases "meter", "m", International metre", "kilometer", and "km", as defined by inla.wkt_unit_params or the params argument. (For legacy PROJ4 use, only "m" and "km" are supported)

params
Length unit definitions, in the list format produced by inla.wkt_unit_params(), Default: NULL, which invokes inla.wkt_unit_params()

Value

For inla.wkt_unit_params, a list of named unit definitions

For inla.wkt_get_lengthunit, a list of length units used in the wkt string, excluding the ellipsoid radius unit.

For inla.wkt_set_lengthunit, a WKT2 string with altered length units. Note that the length unit for the ellipsoid radius is unchanged.

For inla.crs_get_wkt, WKT2 string.

For inla.crs_get_lengthunit, a list of length units used in the wkt string, excluding the ellipsoid radius unit. (For legacy PROJ4 code, the raw units from the proj4string are returned, if present.)

For inla.crs_set_lengthunit, a sp::CRS object with altered length units. Note that the length unit for the ellipsoid radius is unchanged.

For inla.wkt_unit_params, a list of named unit definitions

For inla.wkt_get_lengthunit, a list of length units used in the wkt string, excluding the ellipsoid radius unit.

For inla.wkt_set_lengthunit, a WKT2 string with altered length units. Note that the length unit for the ellipsoid radius is unchanged.

For inla.crs_get_wkt, WKT2 string.

For inla.crs_get_lengthunit, a list of length units used in the wkt string, excluding the ellipsoid radius unit. (For legacy PROJ4 code, the raw units from the proj4string are returned, if present.)

For inla.crs_set_lengthunit, a sp::CRS object with altered length units. Note that the length unit for the ellipsoid radius is unchanged.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.sp_get_crs()

inla.sp_get_crs()

Examples

## Not run: 
if (inla.has_PROJ6()) {
    c1 <- inla.CRS("globe")
    inla.crs_get_lengthunit(c1)
    c2 <- inla.crs_set_lengthunit(c1, "km")
    inla.crs_get_lengthunit(c2)
}

## End(Not run)

## Not run: 
if (inla.has_PROJ6()) {
    c1 <- inla.CRS("globe")
    inla.crs_get_lengthunit(c1)
    c2 <- inla.crs_set_lengthunit(c1, "km")
    inla.crs_get_lengthunit(c2)
}

## End(Not run)

[Package INLA version 22.02.16-2 Index]