fgn {INLA}	R Documentation

Return the coefficients in the 3-component AR(1) mixture representing FGN(H)

Description

This function will return the coefficients in the 3-component AR(1) mixture representing FGN(H)

Usage

     inla.fgn(H, K=4L, lag.max = NULL, approx = TRUE)
 

Arguments

H
The Hurst coeffcient (0<H<1), or a vector of those

K
The number of components in representation, must be 3L or 4L

lag.max
Integer. If positive integer, return the coeffcients implicitely as the ACF from 0 to lag.max

approx
Logical. If lag.max is an positive integer and approx is FALSE, then return the true ACF instead of the approximated one.

Value

inla.fgn returns a named matrix. If is.null(lag.max), then first column is H, columns 1+1:K are lag one correlations (or phi's), and columns 1+K+1:K are the weights. If lag.max > 0, then return the ACFs in columns 2+(0:lag.max), for the H in column 1, either the approximated ones or the the true ones.

This function is EXPERIMENTAL!!!

Author(s)

Havard Rue hrue@r-inla.org

Examples

     r = c(inla.fgn(0.7))
     r_m = inla.fgn(seq(0.6, 0.8, by=0.01))
 

[Package INLA version 22.02.16-2 Index]