geobugs2inla {INLA}	R Documentation

INLA utility functions

Description

Various utility functions for INLA

Usage

     inla.geobugs2inla(adj, num, graph.file="graph.dat")

Arguments

adj
A vector listing the ID numbers of the adjacent areas for each area. This is a sparse representation of the full adjacency matrix for the study region, and can be generated using the Adjacency Tool from the Map menu in GeoBUGS.

num
A vector of length N (the total number of areas) giving the number of neighbours n.i for each area.

graph.file
Name of the file of the new graph in the INLA format.

Value

The return value is the name of the graph-file created.

Note

These are all the same function, and the two different names are due to backward-compatibility

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla, inla.surv, hyperpar.inla

[Package INLA version 22.02.16-2 Index]