Improved estimates for the hyperparameters

Improve the estimates of the posterior marginals for the hyperparameters of the model using the grid integration strategy.

inla.hyperpar(
         result,
         skip.configurations = TRUE,
         verbose = FALSE,
         dz = 0.75,
         diff.logdens = 15,
         h = NULL,
         restart = FALSE,
         quantiles = NULL, 
         keep = FALSE)

Arguments

result

An object of class inla, ie a result of a call to inla()

skip.configurations

A boolean variable; skip configurations if the values at the main axis are to small. (Default TRUE)

verbose

Boolean indicating wheather the inla program should run in a verbose mode.

dz

Step length in the standardized scale used in the construction of the grid, default 0.75.

diff.logdens

The difference of the log.density for the hyperpameters to stop numerical integration using int.strategy='grid'. Default 15

h

The step-length for the gradient calculations for the hyperparameters. Default 0.01.

restart

A boolean defining wheather the optimizer should start again to ind the mode or if it should use the mode contained in the object

quantiles

A vector of quantiles, to compute for each posterior marginal.

keep

A boolean variable indicating the working files (ini file, data files and results files) should be kept

Value

The object returned is the same as object but the estimates of the hyperparameters are replaced by improved estimates.

References

See the references in inla

Author

Havard Rue hrue@r-inla.org

Note

This function might take a long time or if the number of hyperparameters in the model is large. If it complains and says I cannot get enough memory, try to increase the value of the argument dz or decrease diff.logdens.

See also