inla.iidkd.sample {INLA}	R Documentation

Provide samples from the iidkd component (experimental)

Description

This function provide samples of the iidkd component using more interpretable parameters

Usage

inla.iidkd.sample(n = 10^4, result, name, return.cov = FALSE)

Arguments

n
Integer Number of samples to use

result
inla-object An object of class inla, ie a result of a call to inla()

name
Character The name of the iidkd component

return.cov
Logical Return samples of the covariance matrix instead of stdev/correlation matrix described below?

Value

A list of sampled matrices, with (default) correlations on the off-diagonal and standard-deviations on the diagonal

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla.doc("iidkd")

[Package INLA version 22.02.16-2 Index]