Create a coordinate reference system object

Creates either a CRS object or an inla.CRS object, describing a coordinate reference system

inla.CRS(
  projargs = NULL,
  doCheckCRSArgs = TRUE,
  args = NULL,
  oblique = NULL,
  SRS_string = NULL,
  ...
)

inla.wkt_predef()

Arguments

projargs

Either 1) a projection argument string suitable as input to sp::CRS, or 2) an existing CRS object, or 3) a shortcut reference string to a predefined projection; run names(inla.wkt_predef()) for valid predefined projections.

doCheckCRSArgs

default TRUE, must be set to FALSE by package developers including CRS in an S4 class definition to avoid uncontrollable loading of the rgdal namespace.

args

An optional list of name/value pairs to add to and/or override the PROJ4 arguments in projargs. name=value is converted to "+name=value", and name=NA is converted to "+name".

oblique

Vector of length at most 4 of rotation angles (in degrees) for an oblique projection, all values defaulting to zero. The values indicate (longitude, latitude, orientation, orbit), as explained in the Details section below.

SRS_string

a WKT2 string defining the coordinate system; see sp::CRS. This takes precedence over projargs.

...

Additional parameters. Not currently in use.

Value

Either an sp::CRS object or an inla.CRS object, depending on if the coordinate reference system described by the parameters can be expressed with a pure sp::CRS object or not.

An S3 inla.CRS object is a list, usually (but not necessarily) containing at least one element:

crs

The basic sp::CRS object

inla.wkt_predef returns a WKT2 string defining a projection

Details

The first two elements of the oblique vector are the (longitude, latitude) coordinates for the oblique centre point. The third value (orientation) is a counterclockwise rotation angle for an observer looking at the centre point from outside the sphere. The fourth value is the quasi-longitude (orbit angle) for a rotation along the oblique observers equator.

Simple oblique: oblique=c(0, 45)

Polar: oblique=c(0, 90)

Quasi-transversal: oblique=c(0, 0, 90)

Satellite orbit viewpoint: oblique=c(lon0-time*v1, 0, orbitangle, orbit0+time*v2), where lon0 is the longitude at which a satellite orbit crosses the equator at time=0, when the satellite is at an angle orbit0 further along in its orbit. The orbital angle relative to the equatorial plane is orbitangle, and v1 and v2 are the angular velocities of the planet and the satellite, respectively. Note that "forward" from the satellite's point of view is "to the right" in the projection.

When oblique[2] or oblique[3] are non-zero, the resulting projection is only correct for perfect spheres.

See also

Author

Finn Lindgren finn.lindgren@gmail.com

Examples

if (require(rgdal)) {
  crs1 <- inla.CRS("longlat_globe")
  crs2 <- inla.CRS("lambert_globe")
  crs3 <- inla.CRS("mollweide_norm")
  crs4 <- inla.CRS("hammer_globe")
  crs5 <- inla.CRS("sphere")
  crs6 <- inla.CRS("globe")
} 
#> Loading required package: rgdal
#> Warning: there is no package called ‘rgdal’
if (FALSE) { # \dontrun{
 names(inla.wkt_predef())
} # }