Functions for defining the Barrier models

Functions for defining Barrier models as an inla rgeneric model

inla.barrier.pcmatern(
  mesh,
  barrier.triangles,
  prior.range,
  prior.sigma,
  range.fraction = 0.2,
  enable.INLAspacetime = TRUE
)

inla.barrier.polygon(mesh, barrier.triangles, Omega = NULL)

inla.barrier.q(fem, ranges, sigma = 1, envir = NULL)

inla.barrier.fem(mesh, barrier.triangles, Omega = NULL)

Arguments

mesh

The mesh to build the model on, from inla.mesh.2d

barrier.triangles

The numerical ids of the triangles that make up the barrier area

prior.range

2 parameters (range0,Prange) for the prior spatial range. If Prange is NA, then range0 is used as a fixed range value (not tested).

prior.sigma

2 parameters (sig0,Psig) for the prior marginal standard deviation sigma. If Psig is NA, then sig0 is used as a fixed sigma value (not tested).

range.fraction

The length of the spatial range inside the barrier area, as a fraction of the range parameter.

enable.INLAspacetime

Use the implentation in the package INLAspacetime instead if available (default TRUE) if its available. You may need set this option to FALSE if you want to extract properties of the model for other use, like for example inla.rgeneric.q.

Omega

Advanced option for creating a set of permeable barriers (not documented)

fem

represents the Barrier model or the Different Terrains (DT) model, by containing all the needed matrices to solve the SPDE

ranges, sigma

the hyperparameters that determine Q

envir

the environment used for caching (with optimize=TRUE), if any

Value

inla.barrier.pcmatern gives the (rgeneric) model object for fitting the model in INLA

inla.barrier.polygon gives the polygon around the barrier (mainly for plotting)

inla.barrier.q is an internal method producing the Q matrix from a result of inla.barrier.fem,

inla.barrier.fem is an internal method producing the Finite Element matrices.

Details

This model is described in the ArXiv preprint arXiv:1608.03787. For examples, see https://haakonbakkagit.github.io/btopic128.html

inla.barrier.pcmatern This function creates the model component used in inla(...)

inla.barrier.polygon This function constructs SpatialPolygons for the different subdomains (areas)

inla.barrier.q: This function computes a specific precision matrix

inla.barrier.fem This function computes the Finite Element matrices that are needed to compute the precision matrix Q later

See also

Author

Haakon Bakka bakka@r-inla.org