inla.diameter {INLA}	R Documentation

Diameter of a point set

Description

Find an upper bound to the convex hull of a point set

Usage

inla.diameter(x, ...)

## Default S3 method:
inla.diameter(x, manifold = "", ...)

## S3 method for class 'inla.mesh.1d'
inla.diameter(x, ...)

## S3 method for class 'inla.mesh'
inla.diameter(x, ...)

## S3 method for class 'inla.mesh.segment'
inla.diameter(x, ...)

## S3 method for class 'inla.mesh.lattice'
inla.diameter(x, ...)

Arguments

x
A point set as an n x d matrix, or an inla.mesh related object.

...
Additional parameters passed on to other methods.

manifold
Character string specifying the manifold type. Default is to treat the point set with Euclidean R^d metrics. Use manifold="S2" for great circle distances on the unit sphere (this is set automatically for inla.mesh objects).

Details

inla.diameter.default Calculate upper bound for the diameter of a point set, by encapsulating in a circular domain.

Value

A scalar, upper bound for the diameter of the convex hull of the point set.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

Examples

inla.diameter(matrix(c(0, 1, 1, 0, 0, 0, 1, 1), 4, 2))

[Package INLA version 22.02.16-2 Index]