Compute various mesh related quantities.

Low level function for computing finite element matrices, spherical harmonics, B-splines, and point mappings with barycentric triangle coordinates.

inla.fmesher.smorg(loc, tv,
                   fem = NULL, aniso = NULL,
                   gradients = FALSE,
                   sph0 = NULL, sph = NULL, bspline = NULL,
                   points2mesh = NULL,
                   splitlines = NULL,
                   output = NULL,
                   keep = FALSE)

Arguments

loc

3-column triangle vertex coordinate matrix.

tv

3-column triangle vertex index matrix.

fem

Maximum finite element matrix order to be computed.

aniso

A two-element list with \(\gamma\) and \(v\) for an anisotropic operator \(\nabla\cdot H \nabla\), where \(H=\gamma I + v v^\top\)

gradients

When TRUE, calculate derivative operator matrices dx, dy, and dz.

sph0

Maximal order of rotationally invariant spherical harmonics.

sph

Maximal order of general spherical harmonics.

bspline

Rotationally invariant B-splines on a sphere. 3-vector with number of basis functions n, basis degree degree, and a logical; TRUE uniform knot angles, FALSE for uniform spacing in \(\sin(latitude)\).

points2mesh

3-column matrix with points to be located in the mesh.

splitlines

A list with elements loc (3-column coordinate matrix) and idx (2-column index matrix) describing line segments that are to be split into sub-segments at triangle boundaries.

output

Names of objects to be included in the output, if different from defaults.

keep

When TRUE, for debugging purposes keep the fmesher I/O files on disk.

Value

A list of generated named quantities.

Author

Finn Lindgren finn.lindgren@gmail.com