Generate text RGB color specifications.

Generates a tex RGB color specification matrix based on a color palette.

inla.generate.colors(color,
                     color.axis = NULL,
                     color.n = 512,
                     color.palette = cm.colors,
                     color.truncate = FALSE,
                     alpha = NULL)

Arguments

color

character, matrix or vector

color.axis

The min/max limit values for the color mapping.

color.n

The number of colors to use in the color palette.

color.palette

A color palette function.

color.truncate

If TRUE, truncate the colors at the color axis limits.

alpha

Transparency/opaqueness values.

Author

Finn Lindgren finn.lindgren@gmail.com