inla.has_PROJ6 {INLA}	R Documentation

PROJ6 detection

Description

Detect whether PROJ6 is available for INLA

Usage

inla.has_PROJ6()

inla.not_for_PROJ6(fun)

inla.not_for_PROJ4(fun)

inla.fallback_PROJ6(fun)

inla.requires_PROJ6(fun)

Arguments

fun
The name of the calling function

Details

inla.has_PROJ6 is called to check if PROJ6&GDAL3 are available.

Value

For inla.has_PROJ6, logical; TRUE if PROJ6 is available, FALSE otherwise

Functions

inla.not_for_PROJ6: Called to warn about using old PROJ4 features even though PROJ6 is available

inla.not_for_PROJ4: Called to give an error when calling methods that are only available for PROJ6

inla.fallback_PROJ6: Called to warn about falling back to using old PROJ4 methods when a RPOJ6 method hasn't been implemented

inla.requires_PROJ6: Called to give an error when PROJ6 is required but not available

Examples

inla.has_PROJ6()

[Package INLA version 22.02.16-2 Index]