inla.matern.cov {INLA}	R Documentation

Numerical evaluation of Matern and related covariance functions.

Description

Calculates covariance and correlation functions for Matern models and related oscillating SPDE models, on R^d and on the sphere, S^2.

Usage

inla.matern.cov(
  nu,
  kappa,
  x,
  d = 1,
  corr = FALSE,
  norm.corr = FALSE,
  theta,
  epsilon = 1e-08
)

inla.matern.cov.s2(nu, kappa, x, norm.corr = FALSE, theta = 0, freq.max = NULL)

inla.matern.cov.s2.unstable(
  nu,
  kappa,
  x,
  norm.corr = FALSE,
  theta = 0,
  freq.max = NULL
)

Arguments

nu
The Matern smoothness parameter.

kappa
The spatial scale parameter.

x
Distance values.

d
Space dimension; the domain is R^d.

corr
If TRUE, calculate correlations, otherwise calculate covariances. Only used for pure Matern models (i.e. with theta=0).

norm.corr
If TRUE, normalise by the estimated variance, giving approximate correlations.

theta
Oscillation strength parameter.

epsilon
Tolerance for detecting points close to distance zero.

freq.max
The maximum allowed harmonic order. Current default 40, to be changed to a dynamic choice based on error bounds.

Details

On R^d, the models are defined by the spectral density given by

S(w) = 1 / ( (2π)^d * (kappa^4 + 2 kappa^2 * cos(pi * theta) * |w|^2 + |w|^4)^((nu + d/2)/2) )

On S^2, the models are defined by the spectral coefficients

S(k) = (2k+1) / (4 pi (kappa^4 + 2 kappa^2 cos(pi theta) k(k+1) + k^2(k+1)^2)^((ν + 1)/2) )

inla.matern.cov.s2.unstable is an old implementation using orthopolynom::legendre.polynomials and orthopolynom::polynomial.values

Author(s)

Finn Lindgren finn.lindgren@gmail.com

[Package INLA version 22.02.16-2 Index]