inla.mesh.1d {INLA}	R Documentation

Function space definition objects for 1D SPDE models.

Description

Create a 1D mesh specification inla.mesh.1d object, that defines a function space for 1D SPDE models.

Usage

inla.mesh.1d(
  loc,
  interval = range(loc),
  boundary = NULL,
  degree = 1,
  free.clamped = FALSE,
  ...
)

inla.mesh.1d.fem(mesh)

Arguments

loc
B-spline knot locations.

interval
Interval domain endpoints.

boundary
Boundary condition specification. Valid conditions are c('neumann', 'dirichlet', 'free', 'cyclic'). Two separate values can be specified, one applied to each endpoint.

degree
The B-spline basis degree. Supported values are 0, 1, and 2.

free.clamped
If TRUE, for 'free' boundaries, clamp the basis functions to the interval endpoints.

...
Additional option, currently unused.

mesh
An inla.mesh.1d object

Author(s)

Finn Lindgren finn.lindgren@gmail.com

[Package INLA version 22.02.16-2 Index]