inla.mesh.1d.bary {INLA}	R Documentation

Mapping matrix for 1D meshes

Description

Calculates barycentric coordinates and weight matrices for inla.mesh.1d() objects.

Usage

inla.mesh.1d.bary(mesh, loc, method = c("linear", "nearest"))

inla.mesh.1d.A(
  mesh,
  loc,
  weights = NULL,
  derivatives = NULL,
  method = c("linear", "nearest", "quadratic")
)

Arguments

mesh
An inla.mesh.1d() object.

loc
Coordinate values.

method
Interpolation method. If not specified for inla.mesh.1d.A (recommended), it is determined by the mesh basis function properties.

weights
Weights to be applied to the A matrix rows.

derivatives
If TRUE, also compute derivative weight matrices dA and d2A.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

[Package INLA version 22.02.16-2 Index]