inla.mesh.create.helper {INLA}	R Documentation

High-quality triangulations

Description

Create a triangle mesh based on initial point locations, specified or automatic boundaries, and mesh quality parameters.

Usage

inla.mesh.create.helper(points = NULL, points.domain = NULL, ...)

Arguments

points
Matrix of point locations to be used as initial triangulation nodes.

points.domain
Matrix of point locations used to determine the domain extent.

...
Additional arguments passed on to inla.mesh.2d().

Note

Obsolete function kept for backwards compatibility. Use inla.mesh.2d() instead.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.mesh.2d()

[Package INLA version 22.02.16-2 Index]