inla.mesh.lattice {INLA}	R Documentation

Lattice grids for inla.mesh

Description

Construct a lattice grid for inla.mesh()

Usage

inla.mesh.lattice(
  x = seq(0, 1, length.out = 2),
  y = seq(0, 1, length.out = 2),
  z = NULL,
  dims = if (is.matrix(x)) {     dim(x) } else {     c(length(x), length(y)) },
  units = NULL,
  crs = NULL
)

Arguments

x
vector or grid matrix of x-values

y
vector of grid matrix of y-values

z
if x is a matrix, a grid matrix of z-values

dims
the size of the grid, length 2 vector

units
One of c("default", "longlat", "longsinlat").

crs
An optional CRS or inla.CRS object

Value

An inla.mesh.lattice object.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.mesh()

Examples

lattice <- inla.mesh.lattice(seq(0, 1, length.out = 17), seq(0, 1, length.out = 10))

## Use the lattice "as-is", without refinement:
mesh <- inla.mesh.create(lattice = lattice, boundary = lattice$segm)
mesh <- inla.mesh.create(lattice = lattice, extend = FALSE)
plot(mesh)

## Refine the triangulation, with limits on triangle angles and edges:
mesh <- inla.mesh.create(
    lattice = lattice,
    refine = list(max.edge = 0.08),
    extend = FALSE
)
plot(mesh)

## Add an extension around the lattice, but maintain the lattice edges:
mesh <- inla.mesh.create(
    lattice = lattice,
    refine = list(max.edge = 0.08),
    interior = lattice$segm
)
plot(mesh)

## Only add extension:
mesh <- inla.mesh.create(lattice = lattice, refine = list(max.edge = 0.08))
plot(mesh)

[Package INLA version 22.02.16-2 Index]