inla.mesh.map.lim {INLA}	R Documentation

Coordinate mappings for inla.mesh projections.

Description

Calculates coordinate mappings for inla.mesh projections.

Usage

inla.mesh.map.lim(
  loc = NULL,
  projection = c("default", "longlat", "longsinlat", "mollweide")
)

inla.mesh.map(
  loc,
  projection = c("default", "longlat", "longsinlat", "mollweide"),
  inverse = TRUE
)

Arguments

loc
Coordinates to be mapped.

projection
The projection type.

inverse
If TRUE, loc are map coordinates and coordinates in the mesh domain are calculated. If FALSE, loc are coordinates in the mesh domain and the forward map projection is calculated.

Value

For inla.mesh.map.lim, a list:

xlim
X axis limits in the map domain

ylim
Y axis limits in the map domain

No attempt is made to find minimal limits for partial spherical domains.

Functions

inla.mesh.map.lim: Projection extent limit calculations

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.mesh.project()

[Package INLA version 22.02.16-2 Index]