inla.row.kron {INLA}	R Documentation

Row-wise Kronecker products

Description

Takes two Matrices and computes the row-wise Kronecker product. Optionally applies row-wise weights and/or applies an additional 0/1 row-wise Kronecker matrix product, as needed by inla.spde.make.A().

Usage

inla.row.kron(M1, M2, repl = NULL, n.repl = NULL, weights = NULL)

Arguments

M1
A matrix that can be transformed into a sparse Matrix.

M2
A matrix that can be transformed into a sparse Matrix.

repl
An optional index vector. For each entry, specifies which replicate the row belongs to, in the sense used in inla.spde.make.A().

n.repl
The maximum replicate index, in the sense used in inla.spde.make.A().

weights
Optional scaling weights to be applied row-wise to the resulting matrix.

Value

A sparseMatrix object.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.spde.make.A()

[Package INLA version 22.02.16-2 Index]