inla.spde.make.block.A {INLA}	R Documentation

Observation matrices for mesh models.

Description

Constructs observation/prediction weight matrices for numerical integration schemes for regional data problems. Primarily intended for internal use by inla.spde.make.A().

Usage

inla.spde.make.block.A(
  A,
  block,
  n.block = max(block),
  weights = NULL,
  rescale = c("none", "count", "weights", "sum")
)

Arguments

A
A precomputed observation/prediction matrix for locations that are to be joined.

block
Indices specifying block groupings: Entries with the same block value are joined into a single row in the resulting matrix, and the block values are the row indices.

n.block
The number of blocks.

weights
Optional scaling weights to be applied row-wise to the input A matrix.

rescale
Specifies what scaling method should be used when joining the rows of the A matrix as grouped by the block specification.

'none': Straight sum, no rescaling.

'count': Divide by the number of entries in the block.

'weights': Divide by the sum of the weight values within each block.

'sum': Divide by the resulting row sums.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.spde.make.A()

[Package INLA version 22.02.16-2 Index]