inla.spde.models {INLA}	R Documentation

List SPDE models supported by inla.spde objects

Description

List SPDE models supported by inla.spde objects

Usage

inla.spde.models(function.names = FALSE)

inla.spde1.models()

inla.spde2.models()

Arguments

function.names
If FALSE, return list model name lists. If TRUE, return list of model object constructor function names.

Details

Returns a list of available SPDE model type name lists, one for each inla.spde model class (currently inla.spde1() and inla.spde2()).

Value

List of available SPDE model type name lists.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

Examples

## Display help for each supported inla.spde2 model:
for (model in inla.spde2.models()) {
      print(help(paste("inla.spde2.", model, sep = "")))
  }

## Display help for each supported inla.spde* model:
models <- inla.spde.models()
for (type in names(models)) {
      for (model in models[[type]]) {
            print(help(paste("inla.", type, ".", model, sep = "")))
        }
  }

## Display help for each supported inla.spde* model (equivalent to above):
for (model in inla.spde.models(function.names = TRUE)) {
      print(help(model))
  }

[Package INLA version 22.02.16-2 Index]