inla.spde.precision {INLA}	R Documentation

Precision matrices for SPDE models Calculates the precision matrix for given parameter values based on an inla.spde model object.

Description

Precision matrices for SPDE models

Calculates the precision matrix for given parameter values based on an inla.spde model object.

Usage

inla.spde.precision(...)

inla.spde1.precision(spde, ...)

## S3 method for class 'inla.spde1'
inla.spde.precision(spde, ...)

inla.spde2.precision(
  spde,
  theta = NULL,
  phi0 = inla.spde2.theta2phi0(spde, theta),
  phi1 = inla.spde2.theta2phi1(spde, theta),
  phi2 = inla.spde2.theta2phi2(spde, theta),
  ...
)

## S3 method for class 'inla.spde2'
inla.spde.precision(
  spde,
  theta = NULL,
  phi0 = inla.spde2.theta2phi0(spde, theta),
  phi1 = inla.spde2.theta2phi1(spde, theta),
  phi2 = inla.spde2.theta2phi2(spde, theta),
  ...
)

Arguments

...
Additional parameters passed on to other methods.

spde
An inla.spde object.

theta
The parameter vector.

phi0
Internal parameter for a generic model. Expert option only.

phi1
Internal parameter for a generic model. Expert option only.

phi2
Internal parameter for a generic model. Expert option only.

Value

A sparse precision matrix.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.spde.models(), inla.spde2.generic(), inla.spde2.theta2phi0(), inla.spde2.theta2phi1(), inla.spde2.theta2phi2()

[Package INLA version 22.02.16-2 Index]