inla.spde2.matern.sd.basis {INLA}	R Documentation

Approximate variance-compensating basis functions

Description

Calculates an approximate basis for tau and kappa for an inla.spde2.matern model where tau is a rescaling parameter.

Usage

inla.spde2.matern.sd.basis(
  mesh,
  B.sd,
  B.range,
  method = 1,
  local.offset.compensation = FALSE,
  alpha = 2,
  ...
)

Arguments

mesh
An inla.mesh() object.

B.sd
Desired basis for log-standard deviations.

B.range
Desired basis for spatial range.

method
Construction method selector. Expert option only.

local.offset.compensation
If FALSE, only compensate in the average for the tau offset.

alpha
The model alpha parameter.

...
Additional parameters passed on to internal inla.spde2.matern calls.

Value

List of basis specifications

B.tau
Basis for log(tau)

B.kappa
Basis for log(kappa)

Intended for passing on to inla.spde2.matern().

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.spde2.matern()

[Package INLA version 22.02.16-2 Index]