inla.get.inlaEnv {INLA}	R Documentation

Return the internal environment used by INLA

Description

A function which return the internal environment used by INLA

Usage

inla.get.inlaEnv()

Value

This function returns the internal environment used by INLA to keep internal variables.

Author(s)

Havard Rue hrue@r-inla.org

[Package INLA version 22.02.16-2 Index]