jp.define {INLA}	R Documentation

Joint-prior models

Description

A framework for defining joint priors in R

Usage

    inla.jp.define(jp = NULL, ...)

Arguments

jp
The jp-function which returns the joint log-prior as a function of argument theta. There is an optional second argument that is a vector of theta-names. If second argument is not present, argument .theta.desc will be added.

...
Named list of variables that defines the environment of jp

Value

This allows joint priors to be defined in R.

This function is for internal use only.

Author(s)

Havard Rue hrue@r-inla.org

[Package INLA version 22.02.16-2 Index]