link {INLA}	R Documentation

Link functions in INLA

Description

Define link-functions and its inverse

Usage

 inla.link.log(x, inverse=FALSE)
 inla.link.invlog(x, inverse=FALSE)
 inla.link.neglog(x, inverse=FALSE)
 inla.link.invneglog(x, inverse=FALSE)
 inla.link.logit(x, inverse=FALSE)
 inla.link.invlogit(x, inverse=FALSE)
 inla.link.probit(x, inverse=FALSE)
 inla.link.invprobit(x, inverse=FALSE)
 inla.link.cloglog(x, inverse=FALSE)
 inla.link.invcloglog(x, inverse=FALSE)
 inla.link.loglog(x, inverse=FALSE)
 inla.link.invloglog(x, inverse=FALSE)
 inla.link.tan(x, inverse=FALSE)
 inla.link.invtan(x, inverse=FALSE)
 inla.link.cauchit(x, inverse=FALSE)
 inla.link.invcauchit(x, inverse=FALSE)
 inla.link.identity(x, inverse=FALSE)
 inla.link.invidentity(x, inverse=FALSE)
 inla.link.inverse(x, inverse=FALSE)
 inla.link.invinverse(x, inverse=FALSE)
 inla.link.robit(x, df=7, inverse=FALSE)
 inla.link.invrobit(x, df=7, inverse=FALSE)
 inla.link.sn(x, intercept=0.5, skew=0, a=0, inverse=FALSE)
 inla.link.invsn(x, intercept=0.5, skew=0, a=0, inverse=FALSE)
 inla.link.invalid(x, inverse=FALSE)
 inla.link.invinvalid(x, inverse=FALSE)
 inla.link.invqpoisson(x, inverse = FALSE, quantile = 0.5)
 

Arguments

x
The argument. A numeric vector.

df
The degrees of freedom for the Student-t

inverse
Logical. Use the link (inverse=FALSE) or its inverse (inverse=TRUE)

intercept
The quantile level for the intercept in the Skew-Normal link

skew
The skewness in the Skew-Normal. Not both of skew and a can be given.

a
The a-paraeter in the Skew-Normal. Not both of skew and a can be given.

quantile
The quantile level for quantile links

Value

Return the values of the link-function or its inverse.

Note

The inv-functions are redundant, as inla.link.invlog(x) = inla.link.log(x, inverse=TRUE) and so on, but they are simpler to use a arguments to other functions.

Author(s)

Havard Rue hrue@r-inla.org

[Package INLA version 22.02.16-2 Index]