inla.load {INLA}	R Documentation

Load or source a file

Description

Load or source a file: (internal use)

Usage

     inla.load(filename, debug = TRUE)
 

Arguments

filename
The name of the file to be loaded, alternatively, sourced.

debug
Logical. Turn on/off debug information.

Details

Try to load the file into the global environment, if that fail, try to source the file into the global environment.

Value

None

Author(s)

Havard Rue hrue@r-inla.org

[Package INLA version 22.02.16-2 Index]