Interactive mesh building and diagnostics

Assess the finite element approximation errors in a mesh for interactive R sessions. More detailed assessment tools are in meshbuilder().

inla.mesh.assessment(mesh, spatial.range, alpha = 2, dims = c(500, 500))

Arguments

mesh

An inla.mesh

spatial.range

numeric; the spatial range parameter to use for the assessment

alpha

numeric; A valid inla.spde2.pcmatern alpha parameter

dims

2-numeric; the grid size

See also

Author

Finn Lindgren finn.lindgren@gmail.com

Examples

library(fmesher)
bnd <- fm_segm(cbind(
    c(0, 10, 10, 0, 0),
    c(0, 0, 10, 10, 0)
), is.bnd = TRUE)
mesh <- fm_mesh_2d_inla(boundary = bnd, max.edge = 1)
out <- inla.mesh.assessment(mesh, spatial.range = 3, alpha = 2)
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qinv", "-r",     reordering, "-S", smtp, paste0("-t", num.threads), qinv.file,     constr.file, out.file), intern = TRUE): error in running command