meshbuilder {INLA}	R Documentation

Interactive mesh building and diagnostics

Description

Interactively design and build a triangle mesh for use with SPDE models, and assess the finite element approximation errors. The R code needed to recreate the mesh outside the interactive Shiny app is also generated. Spatial objects can be imported from the global workspace.

Usage

meshbuilder()

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.mesh.2d(), inla.mesh.create()

Examples

## Not run: 
meshbuilder()

## End(Not run)

[Package INLA version 22.02.16-2 Index]