nwEngland {INLA}	R Documentation

The New England map

Description

This map is used in association to the Leukemia data from Henderson et al (2003); see source.

Format

A SpatialPolygons object.

Source

This map are used to analyse the Leukaemia dataset from

Henderson, R. and Shimakura, S. and Gorst, D., 2002, Modeling spatial variation in leukemia survival data, JASA, 97, 460, 965–972.

Examples

data(Leuk)
plot(nwEngland)

[Package INLA version 22.02.16-2 Index]