Set and get global options for INLA

Set and get global options for INLA

inla.getOption(
  option = c("inla.call", "inla.arg", "fmesher.call", "fmesher.arg", "num.threads",
    "smtp", "safe", "keep", "verbose", "save.memory", "internal.opt",
    "working.directory", "silent", "debug", "show.warning.graph.file",
    "scale.model.default", "short.summary", "inla.timeout", "fmesher.timeout",
    "inla.mode", "malloc.lib", "fmesher.evolution", "fmesher.evolution.warn",
    "fmesher.evolution.verbosity", "INLAjoint.features", "numa")
)

inla.setOption(...)

Arguments

option

The option to get. If option = NULL then inla.getOption then inla.getOption will return a named list of current values, otherwise, option must be one of

inla.call

The path to the inla-program.

inla.arg

Additional arguments to inla.call

fmesher.call

The path to the fmesher-program

fmesher.arg

Additional arguments to fmesher.call

num.threads

Character string with the number of threads to use as A:B, see ?inla

smtp

Sparse matrix library to use, one of band, taucs (default) or pardiso

safe

Run in safe-mode (ie try to automatically fix convergence errors) (default TRUE)

keep

Keep temporary files?

verbose

Verbose output?

save.memory

Save memory at the cost of (minor) accuracy and computing time?

internal.opt

Do internal online optimisations or not

working.directory

The name of the working directory.

silent

Run the inla-program in a silent mode?

debug

Run the inla-program in a debug mode?

cygwin

The home of the Cygwin installation (default "C:/cygwin") (Remote computing for Windows only) (No longer in use!)

ssh.auth.sock

The ssh bind-adress (value of $SSH_AUTH_SOCK int the Cygwin-shell). (Remote computing for Windows only)

show.warning.graph.file

Give a warning for using the obsolete argument graph.file instead of graph

scale.model.default

The default value of argument scale.model which optionally scale intrinisic models to have generalized unit average variance

short.summary

Use a less verbose output for summary. Useful for Markdown documents.

inla.timeout

The timeout limit, in whole seconds, for calls to the inla binary. Default is 0, meaning no timeout limit. Set to a positive integer to terminate inla calls if they run to long. Fractional seconds are rounded up to the nearest integer. This feature is EXPERIMENTAL and might change at a later stage.

fmesher.timeout

The timeout limit, in whole seconds, for calls to the fmesher binary. Default is 0, meaning no timeout limit. Set to a positive integer to terminate fmesher calls that may enter infinite loops due to special geometry regularity. Fractional seconds are rounded up to the nearest integer.

inla.mode

Which mode to use in INLA? Default is "compact". Other options are "classic" and "twostage".

malloc.lib

Which malloc library to use: "je", "tc", "mi", "compiler" or "default". Option "compiler" use the compiler's implementation. The library is loaded using LD_PRELOAD and similar functionality. Loosely, jemalloc is from Facebook, tcmalloc is from Google and mimalloc is from Microsoft. This option is not available for Windows and not all options might be available for every arch. If malloc.lib is a complete path to an external library, that file will be used instead of one of the supported ones.

fmesher.evolution

Control use of fmesher methods during the transition to a separate fmesher package. Levels of fmesher.evolution:

1L uses the intermediate fm_* methods in fmesher that were already available via inlabru from 2.8.0.

2L (current default) uses the full range of fmesher package methods.

Further levels may be added as the package development progresses.

fmesher.evolution.warn

logical; whether to show warnings about deprecated use of legacy INLA methods with fmesher package replacements. When TRUE, shows deprecation messages for many CRS and mesh related methods, pointing to their fm_* replacements. Default since July 2025 is TRUE.

fmesher.evolution.verbosity

logical or character; at what minimum severity to show warnings about deprecated use of legacy INLA methods with fmesher package replacements. When set to "default" (default), "soft", "warn", or "stop", indicates the minimum warning level used when fmesher.evolution.warn is TRUE.

INLAjoint.features

logical Do not use. By purpose left undocumented

numa

logical Enable NUMA features (Linux only)

...

Option and value, like option=value or option, value; see the Examples

Author

Havard Rue hrue@r-inla.org

Examples

 ## set number of threads
 inla.setOption("num.threads", "4:1")
 ## alternative format
 inla.setOption(num.threads="4:1")
 ## check it
 inla.getOption("num.threads")
#> [1] "4:1"