pc.cor0 {INLA}	R Documentation

Utility functions for the PC prior for correlation in AR(1)

Description

Functions to evaluate, sample, compute quantiles and percentiles of the PC prior for the correlation in the Gaussian AR(1) model where the base-model is zero correlation.

Usage

 inla.pc.rcor0(n, u, alpha, lambda)
 inla.pc.dcor0(cor, u, alpha, lambda, log = FALSE)
 inla.pc.qcor0(p, u, alpha, lambda)
 inla.pc.pcor0(q, u, alpha, lambda)
 

Arguments

n
Number of observations

u
The upper limit (see Details)

alpha
The probability going above the upper limit (see Details)

lambda
The rate parameter (see Details)

cor
Vector of correlations

log
Logical. Return the density in natural or log-scale.

p
Vector of probabilities

q
Vector of quantiles

Details

The statement Prob(|cor| > u) = alpha is used to determine lambda unless lambda is given. Either lambda must be given, or u AND alpha. The density is symmetric around zero.

Value

inla.pc.dcor0 gives the density, inla.pc.pcor0 gives the distribution function, inla.pc.qcor0 gives the quantile function, and inla.pc.rcor0 generates random deviates.

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla.doc("pc.rho0")

Examples

 cor = inla.pc.rcor0(100,  lambda = 1)
 d = inla.pc.dcor0(cor, lambda = 1)
 cor = inla.pc.qcor0(c(0.3, 0.7), u = 0.5, alpha=0.01)
 inla.pc.pcor0(cor, u = 0.5, alpha=0.01)
 

[Package INLA version 22.02.16-2 Index]