pc.gammacount {INLA}	R Documentation

Utility functions for the PC prior for the gammacount likelihood

Description

Functions to evaluate, sample, compute quantiles and percentiles of the PC prior for the gammacount likelihood

Usage

 inla.pc.rgammacount(n, lambda = 1)
 inla.pc.dgammacount(x, lambda = 1, log = FALSE)
 inla.pc.qgammacount(p, lambda = 1)
 inla.pc.pgammacount(q, lambda = 1)
 

Arguments

n
Number of observations

lambda
The rate parameter (see Details)

x
Evaluation points

log
Logical. Return the density in natural or log-scale.

p
Vector of probabilities

q
Vector of quantiles

Details

This gives the PC prior for the gammacount likelihood, which is the PC prior for a in Gamma(a, 1) where Gamma(1, 1) is the base model.

Value

inla.pc.dgammacount gives the density, inla.pc.pgammacount gives the distribution function, inla.pc.qgammacount gives the quantile function, and inla.pc.rgammacount generates random deviates.

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla.doc("pc.gammacount")

Examples

 x = inla.pc.rgammacount(100,  lambda = 1)
 d = inla.pc.dgammacount(x, lambda = 1)
 x = inla.pc.qgammacount(0.5, lambda = 1)
 inla.pc.pgammacount(x, lambda = 1)
 

[Package INLA version 22.02.16-2 Index]