pc.gevtail {INLA}	R Documentation

Utility functions for the PC prior for the tail parameter in the GEV likelihood

Description

Functions to evaluate, sample, compute quantiles and percentiles of the PC prior for the tail parameter in the GEV likelihood

Usage

 inla.pc.rgevtail(n, lambda = 7)
 inla.pc.dgevtail(xi, lambda = 7, log = FALSE)
 inla.pc.qgevtail(p, lambda = 7)
 inla.pc.pgevtail(q, lambda = 7)
 

Arguments

n
Number of observations

lambda
The rate parameter in the PC-prior

xi
Vector of evaluation points, where 1>xi>0.

log
Logical. Return the density in natural or log-scale.

p
Vector of probabilities

q
Vector of quantiles

Details

This gives the PC prior for the tail parameter for the GEV likelihood, where xi=0 is the base model.

Value

inla.pc.dgevtail gives the density, inla.pc.pgevtail gives the distribution function, inla.pc.qgevtail gives the quantile function, and inla.pc.rgevtail generates random deviates.

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla.doc("pc.gevtail")

Examples

 xi = inla.pc.rgevtail(100,  lambda = 7)
 d = inla.pc.dgevtail(xi, lambda = 7)
 xi = inla.pc.qgevtail(0.5, lambda = 7)
 inla.pc.pgevtail(xi, lambda = 7)
 

[Package INLA version 22.02.16-2 Index]