pc.ddof {INLA}	R Documentation

PC-prior for dof in a standarized Student-t

Description

A function to evaluate the PC-prior for the degrees of freedom in a standarized Student-t distribution

Usage

     inla.pc.ddof(dof, lambda, u, alpha, log=FALSE)
 

Arguments

dof
Degrees of freedom

log
Logical. Return the density or the log-density

lambda
The optional value of lambda, instead of defining it implicitely through u and alpha

u
The upper value of dof used to elicitate lambda, Prob(dof < u) = alpha

alpha
The probability alpha used to elicitate lambda

Details

These functions implements the PC-prior for the dof in a standarized Student-t distribution (ie. with unit variance and dof > 2). Either lambda, or u AND alpha must be given. Due the internal tabulation, dof must be larger than 2.0025.

Value

inla.pc.ddof returns the prior density for given dof.

Author(s)

Havard Rue hrue@r-inla.org

[Package INLA version 22.02.16-2 Index]