plot.inla {INLA}	R Documentation

Default INLA plotting

Description

Takes am inla object produced by inla and plot the results

Usage

 ## S3 method for class 'inla'
plot(x,
              plot.fixed.effects = TRUE,
              plot.lincomb = TRUE,
              plot.random.effects = TRUE,
              plot.hyperparameters = TRUE,
              plot.predictor = TRUE,
              plot.q = TRUE,
              plot.cpo = TRUE,
              plot.prior = FALSE,
              single = FALSE,
              postscript = FALSE,
              pdf = FALSE,
              prefix = "inla.plots/figure-",
              intern = FALSE,
              debug = FALSE,
              cex = 1.75,
              ...)
 

Arguments

x
A fitted inla object produced by inla

plot.fixed.effects
Boolean indicating if posterior marginals for the fixed effects in the model should be plotted

plot.lincomb
Boolean indicating if posterior marginals for the linear combinations should be plotted

plot.random.effects
Boolean indicating if posterior mean and quantiles for the random effects in the model should be plotted

plot.hyperparameters
Boolean indicating if posterior marginals for the hyperparameters in the model should be plotted

plot.predictor
Boolean indicating if posterior mean and quantiles for the linear predictor in the model should be plotted

plot.q
Boolean indicating if precision matrix should be displayed

plot.cpo
Boolean indicating if CPO/PIT valuesshould be plotted

plot.prior
Plot also the prior density for the hyperparameters

single
Boolean indicating if there should be more than one plot per page (FALSE) or just one (TRUE)

postscript
Boolean indicating if postscript files should be produced instead

pdf
Boolean indicating if PDF files should be produced instead

prefix
The prefix for the created files. Additional numbering and suffix is added.

intern
Plot also the hyperparameters in its internal scale.

debug
Write some debug information

cex
The cex parameter in par(). If set to NULL or 0, then default values will be used for graphics parameters

...
Additional arguments to postscript(), pdf() or dev.new().

Value

The return value is a list of the files created (if any).

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla

Examples

## Not run: 
result = inla(...)
plot(result)
plot(result, single = TRUE, plot.prior = TRUE)
plot(result, single = TRUE, pdf = TRUE, paper = "a4")
   
## End(Not run)
 

[Package INLA version 22.02.16-2 Index]