Plot CRS and inla.CRS objects

Plot the outline of a CRS or inla.CRS projection, with optional graticules (transformed parallels and meridians) and Tissot indicatrices.

# S3 method for class 'inla.CRS'
plot(x,
     xlim = NULL, ylim = NULL,
     outline = TRUE,
     graticule = c(15, 15, 45),
     tissot = c(30, 30, 30),
     asp = 1, add = FALSE,
     eps=0.05, ...)
# S3 method for class 'CRS'
plot(x,
     xlim = NULL, ylim = NULL,
     outline = TRUE,
     graticule = c(15, 15, 45),
     tissot = c(30, 30, 30),
     asp = 1, add = FALSE,
     eps=0.05, ...)

Arguments

x

A CRS or inla.CRS object.

xlim

Optional x-axis limits.

ylim

Optional y-axis limits.

outline

Logical, if TRUE, draw the outline of the projection.

graticule

Vector of length at most 3, to plot meridians with spacing graticule[1] degrees and parallels with spacing graticule[2] degrees. graticule[3] optionally specifies the spacing above and below the first and last parallel. When graticule[1]==0 no meridians are drawn, and when graticule[2]==0 no parallels are drawn. Use graticule=NULL to skip drawing a graticule.

tissot

Vector of length at most 3, to plot Tissot's indicatrices with spacing tissot[1] degrees and parallels with spacing tissot[2] degrees. tissot[3] specifices a scaling factor. Use tissot=NULL to skip drawing a Tissot's indicatrices.

asp

The aspect ratio for the plot, default 1.

add

If TRUE, add the projecton plot to an existing plot.

eps

Clipping tolerance for rudimentary boundary clipping

...

Additional arguments passed on to the internal calls to plot and lines.

Author

Finn Lindgren <finn.lindgren@gmail.com>

See also

Examples

if (require(rgdal)) {
  oblique <- c(0,45,45,0)
  for (projtype in c("longlat", "lambert", "mollweide", "hammer")) {
    plot(inla.CRS(projtype), main=projtype)
    plot(inla.CRS(projtype, oblique=oblique), main=paste("oblique", projtype))
  }
}
#> Loading required package: rgdal
#> Warning: there is no package called ‘rgdal’