inla.qstat {INLA}	R Documentation

Control and view a remote inla-queue

Description

Control and view a remote inla-queue of submitted jobs

Usage

inla.qget(id, remove = TRUE)
inla.qdel(id)
inla.qstat(id)
inla.qlog(id)
inla.qnuke()
## S3 method for class 'inla.q'
summary(object,...)
## S3 method for class 'inla.q'
print(x,...)

Arguments

id
The job-id which is the output from inla when the job is submitted, the job-number or job-name. For inla.qstat, id is optional and if omitted all the jobs will be listed.

remove
Logical If FALSE, leave the job on the server after retrival, otherwise remove it (default).

x
An inla.q-object which is the output from inla.qstat

object
An inla.q-object which is the output from inla.qstat

...
other arguments.

Details

inla.qstat show job(s) on the server, inla.qget fetch the results (and by default remove the files on the server), inla.qdel removes a job on the server and inla.qnuke remove all jobs on the server. inla.qlog fetches the logfile only.

The recommended procedure is to use r=inla(..., inla.call="submit") and then do r=inla.qget(r) at a later stage. If the job is not finished, then r will not be overwritten and this step can be repeated. The reason for this procedure, is that some information usually stored in the result object does not go through the remote server, hence have to be appended to the results that are retrieved from the server. Hence doing r=inla(..., inla.call="submit") and then later retrive it using r=inla.qget(1), say, then r does not contain all the usual information. All the main results are there, but administrative information which is required to call inla.hyperpar or inla.rerun are not there.

Value

inla.qstat returns an inla.q-object with information about current jobs.

Author(s)

Havard Rue

See Also

inla

Examples

## Not run: 
r = inla(y~1, data = data.frame(y=rnorm(10)), inla.call="submit")
inla.qstat()
r = inla.qget(r, remove=FALSE)
inla.qdel(1)
inla.qnuke()

## End(Not run)

[Package INLA version 22.02.16-2 Index]