qreordering {INLA}	R Documentation

Compute the reordering using the GMRFLib implementation

Description

This function compute the reordering (or find the best reordering) using the GMRFLib implementation

Usage

     inla.qreordering(graph, reordering = inla.reorderings())
 

Arguments

graph
A (inla-)graph object

reordering
The name of the reordering algorithm to be used; either one of the names listed in inla.reorderings(). The default is "auto" which try several reordering algorithm and use the best one for this particular matrix.

Value

inla.qreordering returns a list with the name of the reordering algorithm used or found, the reordering code for the reordering algorithm, the actual reordering and its inverse.

Author(s)

Havard Rue hrue@r-inla.org

Examples

 g = system.file("demodata/germany.graph", package="INLA")
 r = inla.qreordering(g)
 m = inla.graph2matrix(g)
 r = inla.qreordering(m)

[Package INLA version 22.02.16-2 Index]