Solves linear SPD systems

This routine use the GMRFLib implementation to solve linear systems with a SPD matrix.

inla.qsolve(
  Q,
  B,
  reordering = inla.reorderings(),
  method = c("solve", "forward", "backward")
)

Arguments

Q

A SPD matrix, either as a (dense) matrix or sparse-matrix

B

The right hand side matrix, either as a (dense) matrix or sparse-matrix.

reordering

The type of reordering algorithm to be used for TAUCS; either one of the names listed in inla.reorderings() or the output from inla.qreordering(Q). The default is "auto" which try several reordering algorithm and use the best one for this particular matrix (using the TAUCS library).

method

The system to solve, one of "solve", "forward" or "backward". Let Q = L L^T, where L is lower triangular (the Cholesky triangle), then method="solve" solves L L^T X = B or equivalently Q X = B, method="forward" solves L X = B, and method="backward" solves L^T X = B.

Value

inla.qsolve returns a matrix X, which is the solution of Q X = B, L X = B or L^T X = B depending on the value of method.

Author

Havard Rue hrue@r-inla.org

Examples

 n = 10
 nb <- n-1
 QQ = matrix(rnorm(n^2), n, n)
 QQ <- QQ %*% t(QQ)

 Q = inla.as.sparse(QQ)
 B = matrix(rnorm(n*nb), n, nb)

 X = inla.qsolve(Q, B, method = "solve")
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qsolve",     "-r", reordering, "-S", smtp, Qfile, Xfile, Bfile, method),     intern = TRUE): error in running command
 XX = inla.qsolve(Q, B, method = "solve", reordering = inla.qreordering(Q))
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qreordering",     "-r", reordering, "-S", "taucs", g.file), intern = TRUE): error in running command
 print(paste("err solve1", sum(abs( Q %*% X - B))))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'X' not found
 print(paste("err solve2", sum(abs( Q %*% XX - B))))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'XX' not found

 ## the forward and backward solve is tricky, as after permutation and with Q=LL', then L is
 ## lower triangular, but L in the orginal ordering is not lower triangular. if the rhs is iid
 ## noise, this is not important. to control the reordering, then the 'taucs' library must be
 ## used.
 inla.setOption(smtp = 'taucs')

 ## case 1. use the matrix as is, no reordering
 r <- "identity"
 L = t(chol(Q))
 X = inla.qsolve(Q, B, method = "forward", reordering = r)
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qsolve",     "-r", reordering, "-S", smtp, Qfile, Xfile, Bfile, method),     intern = TRUE): error in running command
 XX = inla.qsolve(Q, B, method = "backward", reordering = r)
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qsolve",     "-r", reordering, "-S", smtp, Qfile, Xfile, Bfile, method),     intern = TRUE): error in running command
 print(paste("err forward ", sum(abs(L %*% X - B))))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'X' not found
 print(paste("err backward", sum(abs(t(L) %*% XX - B))))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'XX' not found

 ## case 2. use a reordering from the library
 r <- inla.qreordering(Q)
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qreordering",     "-r", reordering, "-S", "taucs", g.file), intern = TRUE): error in running command
 im <- r$ireordering
#> Error in r$ireordering: $ operator is invalid for atomic vectors
 m <- r$reordering
#> Error in r$reordering: $ operator is invalid for atomic vectors
 print(cbind(idx = 1:n, m, im) )
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'm' not found
 Qr <- Q[im, im]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'i' in selecting a method for function '[': object 'im' not found
 L = t(chol(Qr))[m, m]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 't': error in evaluating the argument 'x' in selecting a method for function 'chol': object 'Qr' not found

 X = inla.qsolve(Q, B, method = "forward", reordering = r)
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qsolve",     "-r", reordering, "-S", smtp, Qfile, Xfile, Bfile, method),     intern = TRUE): error in running command
 XX = inla.qsolve(Q, B, method = "backward", reordering = r)
#> Error in system(paste(shQuote(inla.call.no.remote()), "-s -m qsolve",     "-r", reordering, "-S", smtp, Qfile, Xfile, Bfile, method),     intern = TRUE): error in running command
 print(paste("err forward ", sum(abs( L %*% X - B))))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'X' not found
 print(paste("err backward", sum(abs( t(L) %*% XX - B))))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'print': object 'XX' not found