inla.ssh.copy.id {INLA}	R Documentation

Setup remote computing

Description

Initialize the definition file and print the path to the internal script to transfer ssh-keys

Usage

inla.remote()
inla.ssh.copy.id()

Arguments

None

Value

inla.remote is used once to setup the remote host information file (definition file) in the users home directory; see the FAQ entry on this issue for more information. inla.ssh.copy.id will return the path to the internal script to transfer ssh-keys.

Author(s)

Havard Rue hrue@r-inla.org

Examples

##See the FAQ entry on this issue on r-inla.org.

[Package INLA version 22.02.16-2 Index]