inla.rerun {INLA}	R Documentation

Rerun an analysis

Description

Rerun inla on an inla-object (output from link{inla})

Usage

 inla.rerun(object, plain=FALSE)
 

Arguments

object
An inla-object, ie the output from an inla-call

plain
Logical. If FALSE (default), then make changes in object to improve the performance

Value

This function will take the result in object, and rerun inla again. If plain is FALSE, start the optimization from the mode in object so that we can obtain an improvement the mode for the hyperparameters. Otherwise, start from the same configuration as for object. The returned value is an inla-object.

See Also

inla

Examples

r = inla(y ~ 1,  data = data.frame(y=1:10))
r = inla.rerun(r)

[Package INLA version 22.02.16-2 Index]