inla.hyperpar.sample {INLA}	R Documentation

Produce samples from the approximated joint posterior for the hyperparameters

Description

Produce samples from the approximated joint posterior for the hyperparameters

Usage

inla.hyperpar.sample(n, result, intern=FALSE, improve.marginals = FALSE)

Arguments

n
Integer. Number of samples required.

result
An inla-object, f.ex the output from an inla-call.

intern
Logical. If TRUE then produce samples in the internal scale for the hyperparmater, if FALSE then produce samples in the user-scale. (For example log-precision (intern) and precision (user-scale))

improve.marginals
Logical. If TRUE, then improve the samples taking into account possible better marginal estimates for the hyperparameters in result.

Value

A matrix where each sample is a row. The contents of the column is described in the rownames.

Author(s)

Havard Rue hrue@r-inla.org

Examples

n = 100
r = inla(y ~ 1 + f(idx), data = data.frame(y=rnorm(n), idx = 1:n))
ns = 500
x = inla.hyperpar.sample(ns, r)

rr = inla.hyperpar(r)
xx = inla.hyperpar.sample(ns, rr, improve.marginals=TRUE)

[Package INLA version 22.02.16-2 Index]