summary.inla {INLA}	R Documentation

Summary for a INLA fit

Description

Takes a fitted inla or surv.inla object produced by inla or surv.inla and produces a summary from it.

Usage

## S3 method for class 'inla'
summary(object, digits = 3L, include.lincomb = TRUE, ...)
## S3 method for class 'summary.inla'
print(x, digits = 3L, ...)

Arguments

object
a fitted inla object as produced by inla.

x
a summary.inla object produced by summary.inla

digits
Integer Number of digits

include.lincomb
Logcial Include the summary for the the linear combinations or not

...
other arguments.

Details

Posterior mean and standard deviation (together with quantiles or cdf) are printed for the fixed effects in the model.

For the random effects the function summary() prints the posterior mean and standard deviations for the hyperparameters

If the option short.summary is set to TRUE using inla.setOption, then a less verbose summary variant will be used, which might be more suitable for Markdown documents.

Value

summary.inla returns an object of class summary.inla, a list of components to print.

Author(s)

Sara Martino and Havard Rue

See Also

inla

[Package INLA version 22.02.16-2 Index]