Show the version of the INLA-package

Show the version of the INLA-package

inla.version(what = c("default",
                      "version", 
                      "info", 
                      "hgid",
                      "rinla",
                      "inla",
                      "date",
                      "bdate"))

Arguments

what

What to show version of

Value

inla.version either display the current version information using cat with default or info, or return the version number/information for other spesific requests through the call.

Author

Havard Rue hrue@r-inla.org

Examples

## Summary of all
inla.version()
#> 
#> 
#> 	INLA version ............: 0.0-1467268228
#> 	INLA date ...............: Thu 30 Jun 08:30:28 CEST 2016
#> 	INLA hgid ...............: hgid: 36d9f0f3bfd1  date: Thu Jun 30 08:07:08 2016 +0200
#> 	INLA-program hgid .......: hgid: 36d9f0f3bfd1  date: Thu Jun 30 08:07:08 2016 +0200
#> 	Maintainers .............: Havard Rue <hrue@r-inla.org>
#> 	                         : Finn Lindgren <finn.lindgren@gmail.com>
#> 	                         : Daniel Simpson <dp.simpson@gmail.com>
#> 	                         : Elias Teixeira Krainski <elias.krainski@math.ntnu.no>
#> 	                         : Haakon Bakka <bakka@r-inla.org>
#> 	                         : Andrea Riebler <andrea.riebler@math.ntnu.no>
#> 	                         : Geir-Arne Fuglstad <fulgstad@math.ntnu.no>
#> 	Main web-page ...........: www.r-inla.org
#> 	Download-page ...........: inla.r-inla-download.org
#> 	Email support ...........: help@r-inla.org
#> 	                         : r-inla-discussion-group@googlegroups.com
#> 	Source-code .............: bitbucket.org/hrue/r-inla
#> 
## The building date
inla.version("bdate")
#> [1] "201606300830"