inla.as.wkt_tree.wkt {INLA}	R Documentation

Internal WKT handling

Description

Conversion between WKT and a tree representation

Usage

inla.as.wkt_tree.wkt(x, ...)

inla.as.wkt.wkt_tree(x, pretty = FALSE, ...)

inla.wkt_tree_get_item(x, item, duplicate = 1)

inla.wkt_tree_set_item(x, item_tree, duplicate = 1)

Arguments

x
A WKT2 string, or a wkt_tree list structure

...
Unused

item
character vector with item labels identifying a parameter item entry.

duplicate
For items that have more than one match, duplicate indicates the index number of the desired version. Default: 1

item_tree
An item tree identifying a parameter item entry

[Package INLA version 22.02.16-2 Index]