Zambia {INLA}	R Documentation

Semiparametric regression

Description

Undernutrition of children in each region of Zambia is measured through a score computed on the basis of some anthropometric measures. The data set contains also other infomation about each child.

Format

A data frame with 4847 observations on the following 10 variables.

hazstd

standardised Z score of stunting

bmi

body mass index of the mother

agc

age of the child in months

district

district where the child lives

rcw

mother employment status with categories "working" (1) and "not working" (-1)

edu1

mother's educations status with categories "complete primary but incomplete secondary " (edu1=1), "complete secondary or higher" (edu2=1) and "no education or incomplete primary" (edu1=edu2=-1)

edu2

see above

tpr

locality of the domicile with categories "urban" (1) and "rural" (-1)

sex

gender of the child with categories "male" (1) and "female" (-1)

edu

DO NOT KNOW; check source

Source

BayesX Manual http://www.stat.uni-muenchen.de/~bayesx/bayesx.html

Examples

data(Zambia)

[Package INLA version 22.02.16-2 Index]