inla.binary.install {INLA}	R Documentation

Install alternative binary builds

Description

Install alternative binary builds.

Usage

    inla.binary.install(os = c("CentOS Linux-6", "CentOS Linux-7", "CentOS Linux-8",
                               "CentOS Stream-8",
                               "Fedora-33", "Fedora-34", "Fedora Linux-35",
                               "Manjaro Linux",
                               "Ubuntu-16.04", "Ubuntu-18.04", "Ubuntu-20.04"), 
                        path = NULL, verbose = TRUE, md5.check = TRUE)
 

Arguments

os
If os is given, install binary build for this os. If os is not given, chose os interactively among available builds.

path
character. The install path. If NULL the path is derived from INLA package

verbose
Logical. Verbose output if TRUE

md5.check
Logical. If TRUE, stop if md5-checksum-file is not present or md5-checksum fail. If FALSE, ignore md5-checksum check.

Details

Install a new binary for os unless missing(os), for which the os is chosen interactively among the available builds.

Value

Return TRUE if installation was sucessful and FALSE if not.

Author(s)

Havard Rue hrue@r-inla.org

Examples

   ## Not run: 
     inla.binary.install()
     inla.binary.install(os = "CentOS Linux-7")
     inla.binary.install(os = "CentOS Linux-7",  path = "~/local/bin/inla.binary")
   
## End(Not run)
 

[Package INLA version 22.02.16-2 Index]