control.compute {INLA}	R Documentation

Control variables in control.compute

Description

Control variables in control.compute for use in inla

Usage

inla.set.control.compute.default(...)
control.compute(config, cpo, dic, graph, hyperpar, mlik, openmp.strategy, po, q, return.marginals, return.marginals.predictor, smtp, waic)

Arguments

...
Possible arguments

openmp.strategy
The computational strategy to use: 'small', 'medium', 'large', 'huge', 'default' and 'pardiso'.

hyperpar
A boolean variable if the marginal for the hyperparameters should be computed. Default TRUE.

return.marginals
A boolean variable if the marginals for the latent field should be returned (although it is computed). Default TRUE

return.marginals.predictor
A boolean variable if the marginals for the linear predictor should be returned (although it is computed). Default FALSE

dic
A boolean variable if the DIC-value should be computed. Default FALSE.

mlik
A boolean variable if the marginal likelihood should be computed. Default TRUE.

cpo
A boolean variable if the cross-validated predictive measures (cpo, pit) should be computed (default FALSE)

po
A boolean variable if the predictive ordinate should be computed (default FALSE)

waic
A boolean variable if the Watanabe-Akaike information criteria should be computed (default FALSE)

q
A boolean variable if binary images of the precision matrix, the reordered precision matrix and the Cholesky triangle should be generated. (Default FALSE.)

config
A boolean variable if the internal GMRF approximations be stored. (Default FALSE. EXPERIMENTAL)

smtp
The sparse-matrix solver, one of 'default', 'taucs', 'band' or 'pardiso' (default inla.getoption("smtp")). smtp='pardiso' implies openmp.strategy='pardiso'.

graph
A boolean variable if the graph itself should be returned. (Default FALSE.)

Value

The function control.compute is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.compute.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]