control.family {INLA}	R Documentation

Control variables in control.family

Description

Control variables in control.family for use in inla

Usage

inla.set.control.family.default(...)
control.family(beta.censor.value, cenpoisson.I, control.bgev, control.link, control.mix, control.pom, dummy, fixed, gev.scale.xi, hyper, initial, link, param, prior, sn.shape.max, variant)

Arguments

...
Possible arguments

dummy
A dummy argument that can be used as a workaround

hyper
Definition of the hyperparameters

initial
(OBSOLETE!) Initial value for the hyperparameter(s) of the likelihood in the internal scale.

prior
(OBSOLETE!) The name of the prior distribution(s) for othe hyperparameter(s).

param
(OBSOLETE!) The parameters for the prior distribution

fixed
(OBSOLETE!) Boolean variable(s) to say if the hyperparameter(s) is fixed or random.

link
(OBSOLETE! Use control.link=list(model=) instead.) The link function to use.

sn.shape.max
Maximum value for the shape-parameter for Skew Normal observations (default 5.0)

gev.scale.xi
(Expert option, do not use unless you know what you are doing.) The internal scaling of the shape-parameter for the GEV distribution. (default 0.1)

control.bgev
See ?control.bgev

cenpoisson.I
The censoring interval for the censored Poisson

beta.censor.value
The censor value for the Beta-likelihood (0 <= beta.censor.value < 1/2)

variant
This variable is used to give options for various variants of the likelihood, like chosing different parameterisations for example. See the relevant likelihood documentations for options (does only apply to some likelihoods).

control.mix
See ?control.mix

control.pom
See ?control.pom

control.link
See ?control.link

Value

The function control.family is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.family.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]