control.predictor {INLA}	R Documentation

Control variables in control.predictor

Description

Control variables in control.predictor for use in inla

Usage

inla.set.control.predictor.default(...)
control.predictor(A, cdf, compute, cross, fixed, hyper, initial, link, param, precision, prior, quantiles, vb.correct)

Arguments

...
Possible arguments

hyper
Definition of the hyperparameters.

fixed
(OBSOLETE!) If the precision for the artificial noise is fixed or not (defualt TRUE)

prior
(OBSOLETE!) The prior for the artificial noise

param
(OBSOLETE!) Prior parameters for the artificial noise

initial
(OBSOLETE!) The value of the log precision of the artificial noise

compute
A boolean variable; should the marginals for the linear predictor be computed? (Default FALSE.)

cdf
A list of values to compute the CDF for the linear predictor

quantiles
A list of quantiles to compute for the linear predictor

cross
Cross-sum-to-zero constraints with the linear predictor. All linear predictors with the same level of 'cross' are constrained to have sum zero. Use 'NA' for no contribution. 'Cross' has the same length as the linear predictor (including the 'A' matrix extention). (THIS IS AN EXPERIMENTAL OPTION, CHANGES MAY APPEAR.)

A
The observation matrix (matrix or Matrix::sparseMatrix).

precision
The precision for eta* - A*eta, (default exp(15))

link
Define the family-connection for unobserved observations (NA). link is integer values which defines the family connection; family[link[idx]] unless is.na(link[idx]) for which the identity-link is used. The link-argument only influence the fitted.values in the result-object. If is.null(link) (default) then the identity-link is used for all missing observations. If the length of link is 1, then this value is replicated with the length of the responce vector. If an element of the responce vector is !NA then the corresponding entry in link is not used (but must still be a legal value). Setting this variable implies compute=TRUE.

vb.correct
Logical Add this model component to the list of variables to be used for the vb corrected Laplace approximation? If NULL use default choice, otherwise correct if TRUE and do not if FALSE (expert option)

Value

The function control.predictor is used to TAB-complete arguments and returns a list of given arguments. The function inla.set.control.predictor.default returns a list with all the default values of all parameters within this control statement.

See Also

control.update, control.lincomb, control.group, control.mix, control.pom, control.link, control.expert, control.compute, control.lp.scale, control.pardiso.default, control.bgev.default, control.family, control.fixed, control.inla, control.predictor, control.mode, control.hazard, inla

[Package INLA version 22.02.16-2 Index]