inla.cpo {INLA}	R Documentation

Improved estimates for the CPO/PIT-values

Description

Improve the estimates of the CPO/PIT-values be recomputing the model-fit by removing data-points.

Usage

inla.cpo(result,
         force = FALSE,
         mc.cores = NULL,
         verbose = TRUE,
         recompute.mode = TRUE)

Arguments

result
An object of class inla, ie a result of a call to inla()

force
If TRUE, then recompute all CPO/PIT values and not just those with result$cpo$failure > 0.

mc.cores
The number of cores to use in parallel::mclapply. If is.null(mc.cores), then check getOption("mc.cores") and inla.getOption("num.threads") in that order.

verbose
Run in verbose mode?

recompute.mode
Should be mode (and the integration points) be recomputed when a data-point is removed or not?

Value

The object returned is the same as result but the new improved estimates of the CPO/PIT values replaced.

Author(s)

Havard Rue hrue@r-inla.org

See Also

inla

Examples

n = 10
y = rnorm(n)
r = inla(y ~ 1, data = data.frame(y), control.compute = list(cpo=TRUE))

rr = inla.cpo(r, force=TRUE)

[Package INLA version 22.02.16-2 Index]