cut {INLA}	R Documentation

Group-wise model criticism using node-splitting

Description

This function performs group-wise, cross-validatory model assessment for an INLA model using so-called node-splitting (Marshall and Spiegelhalter, 2007; Presanis et al, 2013). The user inputs an object of class inla (i.e. a result of a call to inla()) as well as a variable name (split.by) specifying a grouping: Data points that share the same value of split.by are in the same group. The function then checks whether each group is an "outlier", or in conflict with the remaining groups, using the methodology described in Ferkingstad et al (2017). The result is a vector containing a p-value for each group, corresponding to a test for each group i, where the null hypothesis is that group i is consistent with the other groups except i (so a small p-value is evidence that the group is an "outlier"). See Ferkingstad et al (2017) for further details.

Usage

     inla.cut(result, split.by, mc.cores = NULL, debug=FALSE)
 

Arguments

result
An object of class inla, i.e. a result of a call to inla()

split.by
The name of the variable to group by. Data points that have the same value of split.by are in the same group.

mc.cores
The number of cores to use in parallel::mclapply. If is.null(mc.cores), then check getOption("mc.cores") and inla.getOption("num.threads") in that order.

debug
Print debugging information if TRUE, default is FALSE

Value

A numeric vector of p-values, corresponding to a test for each group i where the null hypothesis is that group i is consistent with the other groups except i. A small p-value for a group indicates that the group is an "outlier" (in conflict with remaining groups).

This function is EXPERIMENTAL!!!

Author(s)

Egil Ferkingstad egil.ferkingstad@gmail.com and Havard Rue hrue@r-inla.org

References

Ferkingstad, E., Held, L. and Rue, H. (2017). Fast and accurate Bayesian model criticism and conflict diagnostics using R-INLA. arXiv preprint arXiv:1708.03272, available at http://arxiv.org/abs/1708.03272. Published in Stat, 6:331-344 (2017).

Marshall, E. C. and Spiegelhalter, D. J. (2007). Identifying outliers in Bayesian hierarchical models: a simulation-based approach. Bayesian Analysis, 2(2):409-444.

Presanis, A. M., Ohlssen, D., Spiegelhalter, D. J., De Angelis, D., et al. (2013). Conflict diagnostics in directed acyclic graphs, with applications in Bayesian evidence synthesis. Statistical Science, 28(3):376-397.

Examples

 ## See http://www.r-inla.org/examples/case-studies/ferkingstad-2017 and Ferkingstad et al (2017).
 

[Package INLA version 22.02.16-2 Index]