inla.mesh.2d {INLA}	R Documentation

High-quality triangulations

Description

Create a triangle mesh based on initial point locations, specified or automatic boundaries, and mesh quality parameters.

Usage

inla.mesh.2d(
  loc = NULL,
  loc.domain = NULL,
  offset = NULL,
  n = NULL,
  boundary = NULL,
  interior = NULL,
  max.edge = NULL,
  min.angle = NULL,
  cutoff = 1e-12,
  max.n.strict = NULL,
  max.n = NULL,
  plot.delay = NULL,
  crs = NULL
)

Arguments

loc
Matrix of point locations to be used as initial triangulation nodes. Can alternatively be a SpatialPoints or SpatialPointsDataFrame object.

loc.domain
Matrix of point locations used to determine the domain extent. Can alternatively be a SpatialPoints or SpatialPointsDataFrame object.

offset
The automatic extension distance. One or two values, for an inner and an optional outer extension. If negative, interpreted as a factor relative to the approximate data diameter (default=-0.10???)

n
The number of initial nodes in the automatic extensions (default=16)

boundary
A list of one or two inla.mesh.segment() objects describing domain boundaries.

interior
An inla.mesh.segment() object describing desired interior edges.

max.edge
The largest allowed triangle edge length. One or two values.

min.angle
The smallest allowed triangle angle. One or two values. (Default=21)

cutoff
The minimum allowed distance between points. Point at most as far apart as this are replaced by a single vertex prior to the mesh refinement step.

max.n.strict
The maximum number of vertices allowed, overriding min.angle and max.edge (default=-1, meaning no limit). One or two values, where the second value gives the number of additional vertices allowed for the extension.

max.n
The maximum number of vertices allowed, overriding max.edge only (default=-1, meaning no limit). One or two values, where the second value gives the number of additional vertices allowed for the extension.

plot.delay
On Linux (and Mac if appropriate X11 libraries are installed), specifying a nonnegative numeric value activates a rudimentary plotting system in the underlying fmesher program, showing the triangulation algorithm at work, with waiting time factor plot.delay between each step.

On all systems, specifying any negative value activates displaying the result after each step of the multi-step domain extension algorithm.

crs
An optional CRS or inla.CRS object

Value

An inla.mesh object.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.mesh.create(), inla.delaunay(), inla.nonconvex.hull()

Examples

loc <- matrix(runif(10 * 2), 10, 2)

if (require("splancs")) {
    boundary <- list(
        inla.nonconvex.hull(loc, 0.1, 0.15),
        inla.nonconvex.hull(loc, 0.2, 0.2)
    )
    offset <- NULL
} else {
    boundary <- NULL
    offset <- c(0.1, 0.2)
}
mesh <- inla.mesh.2d(loc, boundary = boundary, offset = offset, max.edge = c(0.05, 0.1))

plot(mesh)

[Package INLA version 22.02.16-2 Index]