inla.option {INLA}	R Documentation

Set and get global options for INLA

Description

Set and get global options for INLA

Usage

 inla.setOption(...)
 inla.getOption(option)
 

Arguments

...
Option and value, like option=value or option, value; see the Examples

option
The option to get. If option = NULL then inla.getOption then inla.getOption will return a named list of current values, otherwise, option must be one of

inla.call: The path to the inla-program.

inla.arg: Additional arguments to inla.call

fmesher.call: The path to the fmesher-program

fmesher.arg: Additional arguments to fmesher.call

num.threads: Character string with the number of threads to use as A:B

blas.num.threads: Number of threads to use for openblas and mklblas (see inla for details)

smtp: Sparse matrix library to use, one of band, taucs (default) or pardiso

mkl: Use binaries buildt with Intel MKL? (If possible)

vecLib: This option applies to Mac only. If TRUE and mkl=FALSE, link with vecLib BLAS and LAPACK libs (if available)

vecLibPath: This option applies to Mac only. Path to vecLib-libraries. If empty, use default.

pardiso.license: The full path to the PARDISO license file or a newline-separated string with license key(s)

keep: Keep temporary files?

working.directory: The name of the working directory.

silent: Run the inla-program in a silent mode?

debug : Run the inla-program in a debug mode?

cygwin : The home of the Cygwin installation (default "C:/cygwin") [Remote computing for Windows only]

ssh.auth.sock: The ssh bind-adress (value of $SSH_AUTH_SOCK int the Cygwin-shell). [Remote computing for Windows only]

show.warning.graph.file : Give a warning for using the obsolete argument graph.file instead of graph

scale.model.default : The default value of argument scale.model which optionally scale intrinisic models to have generalized unit average variance

short.summary : Use a less verbose output for summary. Useful for Markdown documents.

inla.timeout : The timeout limit, in whole seconds, for calls to the inla binary. Default is 0, meaning no timeout limit. Set to a positive integer to terminate inla calls if they run to long. Fractional seconds are rounded up to the nearest integer. This feature is EXPERIMENTAL and might change at a later stage.

fmesher.timeout : The timeout limit, in whole seconds, for calls to the fmesher binary. Default is 0, meaning no timeout limit. Set to a positive integer to terminate fmesher calls that may enter infinite loops due to special geometry regularity. Fractional seconds are rounded up to the nearest integer.

inla.mode : Which mode to use in INLA? Default is "classic". Other options are (USE AT YOUR OWN RISK, THIS IS WORK IN PROGRESS!!!) options are "twostage" and "experimental".

Author(s)

Havard Rue hrue@r-inla.org

Examples

 ## set number of threads
 inla.setOption("num.threads", "4:1")
 ## alternative format
 inla.setOption(num.threads="4:1")
 ## check it
 inla.getOption("num.threads")

[Package INLA version 22.02.16-2 Index]