Set and get global options for INLA

Set and get global options for INLA

inla.setOption(...)
 inla.getOption(option)

Arguments

...

Option and value, like option=value or option, value; see the Examples

option

The option to get. If option = NULL then inla.getOption then inla.getOption will return a named list of current values, otherwise, option must be one of

inla.call: The path to the inla-program.

inla.arg: Additional arguments to inla.call

fmesher.call: The path to the fmesher-program

fmesher.arg: Additional arguments to fmesher.call

num.threads: Number of threads to use.

blas.num.threads: Number of threads to use for openblas and mklblas (see inla for details)

smtp: Sparse matrix library to use, one of band, taucs (default) or pardiso

mkl: Use binaries buildt with Intel MKL? (If possible)

pardiso.license: The full path to the PARDISO license file

keep: Keep temporary files?

working.directory: The name of the working directory.

silent: Run the inla-program in a silent mode?

debug : Run the inla-program in a debug mode?

internal.binary.mode : if FALSE the (some) output are in ascii format instead of binary format. Using this option, then inla.collect.results will fail (Expert mode)

internal.experimental.mode : Expert option

cygwin : The home of the Cygwin installation (default "C:/cygwin") [Remote computing for Windows only]

ssh.auth.sock: The ssh bind-adress (value of $SSH_AUTH_SOCK int the Cygwin-shell). [Remote computing for Windows only]

enable.inla.argument.weights : if TRUE the inla accepts argument weights

show.warning.graph.file : Give a warning for using the obsolete argument graph.file instead of graph

scale.model.default : The default value of argument scale.model which optionally scale intrinisic models to have generalized unit average variance

short.summary : Use a less verbose output for summary. Useful for Markdown documents.

Author

Havard Rue hrue@r-inla.org

Examples

 ## set number of threads
 inla.setOption("num.threads", 2)
 ## alternative format
 inla.setOption(num.threads=2)
 ## check it
 inla.getOption("num.threads")
#> [1] 2