param2.matern.orig {INLA}	R Documentation

Parameter settings for inla.spde2.matern models.

Description

Construct parameter settings for inla.spde2.matern models.

Usage

param2.matern.orig(
  mesh,
  alpha = 2,
  B.tau = matrix(c(0, 1, 0), 1, 3),
  B.kappa = matrix(c(0, 0, 1), 1, 3),
  prior.variance.nominal = 1,
  prior.range.nominal = NULL,
  prior.tau = NULL,
  prior.kappa = NULL,
  theta.prior.mean = NULL,
  theta.prior.prec = 0.1
)

Arguments

mesh
The mesh to build the model on, as an inla.mesh() object.

alpha
Fractional operator order, 0<α≤q 2 supported. (ν=α-d/2)

B.tau
Matrix with specification of log-linear model for τ.

B.kappa
Matrix with specification of log-linear model for κ.

prior.variance.nominal
Nominal prior mean for the field variance

prior.range.nominal
Nominal prior mean for the spatial range

prior.tau
Prior mean for tau (overrides prior.variance.nominal)

prior.kappa
Prior mean for kappa (overrides prior.range.nominal)

theta.prior.mean
(overrides prior.*)

theta.prior.prec
Scalar, vector or matrix, specifying the joint prior precision for theta.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

inla.spde2.matern()

[Package INLA version 22.02.16-2 Index]