plot.inla.trimesh {INLA}	R Documentation

Low level triangulation mesh plotting

Description

Plots a triangulation mesh using rgl.

Usage

## S3 method for class 'inla.trimesh'
plot(
  x,
  S,
  color = NULL,
  color.axis = NULL,
  color.n = 512,
  color.palette = cm.colors,
  color.truncate = FALSE,
  alpha = NULL,
  lwd = 1,
  specular = "black",
  draw.vertices = TRUE,
  draw.edges = TRUE,
  edge.color = rgb(0.3, 0.3, 0.3),
  ...
)

Arguments

x
A 3-column triangle-to-vertex index map matrix.

S
A 3-column vertex coordinate matrix.

color
Color specification. A single named color, a vector of scalar values, or a matrix of RGB values.

color.axis
The min/max limit values for the color mapping.

color.n
The number of colors to use in the color palette.

color.palette
A color palette function.

color.truncate
If TRUE, truncate the colors at the color axis limits.

alpha
Transparency/opaqueness values. See rgl.material.

lwd
Line width for edges. See rgl.material.

specular
Specular color. See rgl.material.

draw.vertices
If TRUE, draw triangle vertices.

draw.edges
If TRUE, draw triangle edges.

edge.color
Edge color specification.

...
Additional parameters passed to and from other methods.

Author(s)

Finn Lindgren finn.lindgren@gmail.com

See Also

plot.inla.mesh()

[Package INLA version 22.02.16-2 Index]