inla.priors.used {INLA}	R Documentation

Print priors used

Description

Print the priors used for the hyperparameters

Usage

    inla.priors.used(result, digits=6L)

Arguments

result
An inla-object, typically the output from an inla()-call

digits
The digits argument to the function format()

Details

This function provides a more human-friendly output of result$all.hyper of all the priors used for the hyperparameters. Since not all information about the model is encoded in this object, more hyperparameters than actually used, may be printed. In particular, group.theta1 is printed even though the argument group in f() is not used. Similarly for spde-models, but the user should know that, for example, only the two first ones are actually used. Hopefully, this issue will be fixed in the future.

Author(s)

Havard Rue hrue@r-inla.org

Examples

r = inla(y ~ 1 + x, data = data.frame(y = 1:10, x = rep(1:5, 2)))
inla.priors.used(r)

[Package INLA version 22.02.16-2 Index]