inla.version {INLA}	R Documentation

Show the version of the INLA-package

Description

Show the version of the INLA-package

Usage

inla.version(what = c("default", "version", "date"))

Arguments

what
What to show version of

Value

inla.version display the current version information using cat with default or info, or return other spesific requests through the call.

Author(s)

Havard Rue hrue@r-inla.org

Examples

## Summary of all
inla.version()
## The building date
inla.version("date")

[Package INLA version 22.02.16-2 Index]